### 2025-11-04 #9227 @crvernon
:wave: @Stat-Cook - Thanks for your submission to JOSS!  I am going to run this submission by our larger editorial board to ensure that it meets our substantial scholarly effort requirements.  I'll be back in touch ASAP.
