### 2026-01-05 #9728 @editorialbot
[REVIEW]: PID.POS: An R package for the detection of personally identifiable data
**Submitting author:** @Stat-Cook (<a href="http://orcid.org/0000-0003-3343-8271">Robert Cook</a>)
**Repository:** https://github.com/Stat-Cook/pid.pos
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
