### 2026-01-05 Stat-Cook/pid.pos#5 @Stat-Cook
Pre-review
Pre-review notes for [9728](https://github.com/openjournals/joss-reviews/issues/9728):

* Reduce length of paper to be ~1000 words
* Remove redundant license and check its [OSI approved](https://opensource.org/licenses).
* Add [Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) for the package
