### 2026-01-05 57e260d
Add community guidelines to readme
