+++ README.md
@@ -20,3 +20,8 @@ The point of entry for most users will be the `report_on_folder` function.  This
 
 The package is built on the [UDPipe](https://lindat.mff.cuni.cz/services/udpipe/) tree banks via the `udpipe` R package.  Fetching the appropriate udpipe model is intended to be automatic, however if there are issues it is worth looking at the functions `browse_udpipe_repo` and `browse_model_location`.
 
+## Getting help
+
+If you encounter a bug, please file a minimal reproducible example on github. Any feature requests, or requests for guidance on using the package can also be submitted this way though please read the documentation first.
+
+Contribution to pid.pos is welcome, to do so please open an issue so the development team can check it is compatible with the design principles or feel free to fork the repository.
