+++ articles/PaperVignette.md
@@ -1,0 +1,244 @@
+# PID.POS: An R package for the detection of personally identifiable data
+
+## Summary
+
+The `PID.POS` package is designed to aid with the identification of
+personal identifiability risks in data sets. By applying existing
+natural language processing techniques, the package is able to identify
+proper nouns within a data set. The extraction of proper nouns reduced
+the complexity of the data, allowing for a quicker review and oversight
+of the data. The package also includes a basic tool for the design, and
+implementation of a redaction process.
+
+## Statement of need
+
+The world is embedded in a data revolution. Never before have we had the
+depth or breadth of data being captured and analysed than we do at
+present, and this is only set to increase. In response, international
+bodies are taking steps to ensure legal protection of an individual’s
+rights to their own data (European Parliament and Council of the
+European Union 2016). One effect of increase legislation in the European
+Union has been a growing awareness of the role and responsibility of
+data controllers (ICO, n.d.b) and the risks of big data (Clarke 2016).
+Among these concerns, a risk of ‘personal identifiability’ i.e. the
+ability to directly or indirectly identify an individual from a dataset
+(Finck and Pallas 2020), is paramount and, if breeched, can lead to
+reputation damage and fines (ICO, n.d.a).
+
+Where data is structured and comprises only a few hundred observations,
+a manual inspection can identify variables which contain directly
+personally identifiable data (PID) with a reasonable time investment.
+However, in the case of modern large data sets which may comprise
+millions of observations, a manual inspection may miss PID if it is
+embedded within a passage of text, or is a rarity for the given
+variable. The `PID.POS` (Personal Identifiability Detection by Part Of
+Speech tagging) package is designed to aid with the identification of
+PID risks in data sets. In comparison to existing packages which rely on
+a curated list of common names and string-matching, `PID.POS` builds on
+the existing `udpipe` framework(Straka, Hajic, and Straková 2016;
+Wijffels 2023), extracting all examples of proper nouns and providing a
+mechanism for the review and redaction of PID risks.
+
+## Comparison to existing R packages
+
+The need to review data sets to identify risks is not new, and there are
+a number of packages which have been developed to aid in this process.
+The most notable of these are the `PII` package (Patterson-Stein 2025),
+which is designed to identify personally identifiable features via
+pattern matching. These approaches can be effective in identifying PID,
+but have a risk of missing edge cases e.g. relying on sentence case to
+identify names. The approach taken in `PID.POS` conversely takes the
+approach of purposefully extracting all proper-nouns, and hence increase
+the false positive rate, with the intention of supplying a simplified
+extract to aid human interpretation rather than fully automate it.
+
+## In practice
+
+To install the current version of `PID.POS` package, use the following
+code:
+
+``` r
+# install.packages("pak")
+pak::pkg_install("Stat-Cook/PID.POS")
+```
+
+To assist with understanding the `PID.POS` package, we include a subset
+of the ‘friends’ data set from the `friends` package.
+
+``` r
+library(pid.pos)
+the_one_in_massapequa
+```
+
+    #> # A tibble: 257 × 4
+    #>   scene utterance speaker          text                                         
+    #>   <int>     <int> <chr>            <chr>                                        
[…]
