+++ reference/redaction_function_factory.md
@@ -1,0 +1,15 @@
+# Create a redaction function from a \`data.frame\` of replacement rules.
+
+Create a redaction function from a \`data.frame\` of replacement rules.
+
+## Usage
+
+``` r
+redaction_function_factory(rules.frm)
+```
+
+## Arguments
+
+- rules.frm:
+
+  A data.frame with columns \`If\`, \`From\` and \`To\`.
