+++ reference/find_files.md
@@ -1,0 +1,27 @@
+# List all files in a directory with a given extension
+
+List all files in a directory with a given extension
+
+## Usage
+
+``` r
+find_files(path, extensions = get_implemented_extensions(), .lis = list())
+```
+
+## Arguments
+
+- path:
+
+  The path to search for files
+
+- extensions:
+
+  A character vector of file extensions to search for
+
+- .lis:
+
+  A list of file paths
+
+## Value
+
+A list of file paths
