+++ reference/longer_than.md
@@ -1,0 +1,16 @@
+# Produces a function that checks if any strings in a vector is longer than 'limit'
+
+Produces a function that checks if any strings in a vector is longer
+than 'limit'
+
+## Usage
+
+``` r
+longer_than(limit = 5)
+```
+
+## Arguments
+
+- limit:
+
+  The length to check against
