+++ articles/index.md
@@ -1,0 +1,12 @@
+# Articles
+
+### Articles
+
+- [PID.POS: An R package for the detection of personally identifiable
+  data](https://stat-cook.github.io/pid.pos/articles/PaperVignette.md):
+- [Trouble shooting the UDpipe
+  Model](https://stat-cook.github.io/pid.pos/articles/udpipe-model.md):
+- [Reporting on every file in a
+  folder](https://stat-cook.github.io/pid.pos/articles/FolderReport.md):
+- [Automatic Replacement
+  Tools](https://stat-cook.github.io/pid.pos/articles/AutoReplacement.md):
