+++ reference/set_context_window.md
@@ -1,0 +1,16 @@
+# Set the context window size for the \`get_context\` function.
+
+Set the context window size for the \`get_context\` function.
+
+## Usage
+
+``` r
+set_context_window(x)
+```
+
+## Arguments
+
+- x:
+
+  An integer specifying the number of characters to include before and
+  after the token in the context.

[…]
