### 2026-01-06 #9728 @agerada
- [x] **Repository:** Is the source code for this software available at https://github.com/Stat-Cook/pid.pos?
- [ ] **Contribution and authorship:** Has the submitting author ({{author-handle}}) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
