### 2026-01-06 Stat-Cook/pid.pos#6 @agerada
JOSS Review
Hi @Stat-Cook. I am going to write my review comments here, as per https://github.com/openjournals/joss-reviews/issues/9728; I may open separate sub-issues where appropriate.

I'm mainly going to review the software here (paper will be done separately).

## Package documentation.

1. While most functions and datasets seem to be documented, the level of documentation is quite light, making it difficult to use the package without extensively reading the vignettes. For example, one of the core functions (from the point of user interaction with the package) is [`data_frame_report`](https://stat-cook.github.io/pid.pos/reference/data_frame_report.html). The most important part of R function documentation is the examples section, and this is very short (and not executed). I would suggest providing and end-to-end example within this function (and any other "core" functions), using "See also" to cross-reference other functions used in the example.
2. Speaking of cross-references, these do not seem to be used in the package (for example, see [`auto_replace`](https://stat-cook.github.io/pid.pos/reference/auto_replace.html)), which makes navigation a little bit slower. Please could you add these were possible?
3. Argument documentation is also very brief and not sufficient to understand the functions. For example, I wasn't sure what `to_remove` (in `data_frame_report`) does. I think it identifies columns to be excluded from proper noun detection, but this is only after trying out different values for it. The example also does not clarify:

```R
example.data <- head(the_one_in_massapequa, 10)
report1 <- data_frame_report(example.data, to_remove=c("scene", "utterance"))
report2 <- data_frame_report(example.data)
all.equal(report1, report2)
```

4. The "Value" heading states that a tibble is returned. However, the returned object is also a `pid_report` S3 object, which is important to clarify as it there are methods which act on it (e.g., `summary.pid_report`). Use of these methods should also be demonstrated in the example (otherwise it is unlikely that users would discover them).
5. The examples do not clearly indicate data. Suggest change:

```
example.data <- head(the_one_in_massapequa, 10)
data_frame_report(example.data, to_remove=c("scene", "utterance"))
```

to:

```
data(the_one_in_massapequa)
example.data <- head(the_one_in_massapequa, 10)
data_frame_report(example.data, to_remove=c("scene", "utterance"))
```

## Model

The package seems to be primarily a wrapper around `UDPipe`, mainly using a filter for proper noun tokens (i.e., where `upos == "PROPN"). Similar results can be obtained by following: https://cran.r-project.org/web/packages/udpipe/vignettes/udpipe-annotation.html and filtering for "PROPN". Please correct me if that is not the case. In practice, this approach gives many false positive when I test it on a dataset that has already had identifiers removed. Although this is to be expected (and a cautious approach is preferred in such situations), I also see single letters such as "A" labelled as proper nouns. To my eye the approach seems very senstive to capitalisation. I am not too familiar with the model used, but I wonder whether the authors considered other approaches or models that could reduce the false positive rate?

## Error handling/argument validation.

1. Going back to `data_frame_report`, `to_remove` should at least raise a warning if the names are not in the dataframe. Currently,

```R
example.data <- head(the_one_in_massapequa, 10)
isFALSE("abcde" %in% names(example.data))
data_frame_report(example.data, to_remove=c("abcde"))
```

runs without raising any exceptions or warnings.

2. Exceptions should be cleaner and handled earlier. For example, if the user provides an inappropriate `replacement_size` in `random_replacement.f`, the error is cryptic and raised deep rather than handled early (and consequently, uses `size` as argument name rather than `replacement_size`).

## Naming conventions

The package uses `.` in its name and frequently within function names. This is discouraged as it can be confusing due to the S3 dispatch system in R, which uses the `.` convention.

## Features

1. Reversal of replacements. Although this can be done maually, I recommend implementing a convenience reversal function that takes a replacement rules object and safely reverses the replacements.

2. Simple censoring function. A simple censoring function factory that just removes proper nouns, or replaces them with a fixed string, would be useful. It may be possible to do this using the current factories; if so, an example would suffice.

## User extension

I would suggest a vignette that explains how users can create their own function factory. For users unfamiliar with functional programming, this will be non-trivial. Perhaps this would be a good location for a simple censoring function.

## CRAN

Are there plans to submit this package to CRAN? `devtools::check()` only returns some notes which should be easy to address. I think availability on CRAN is important for this package as the package is intended for use on sensitive data, so it this step would give additional assurances to data controllers.

## Other comments

1. The description file contains placeholder text and reference to function that is not in the package (`proper_noun_report`)
2. On loading the package, it saves data to a system location (on a Mac, to Application Support). This is quite unusual for an R package, and I think it would be far safer to explicitly download the model (as in: https://cran.r-project.org/web/packages/udpipe/vignettes/udpipe-annotation.html) or at least bind it temporarily to the R session. If not provided as an argument to functions, they can fall back to downloading it to a temporary location.

---

I shall pause for now and will review `report_on_folder` and the paper itself later – I'll leave this here for now to give time to address these points.
