### 2026-01-06 9e1375a
Warn if column can't be ignored
Add optional check if values given in to_remove occur in the given data set.
