### 2026-01-06 9c1421d
Refactor to_remove as to_ignore
Improve clarity of the to_remove  argument (both docstring and name)
