+++ R/data_frame_report.R
@@ -1,6 +1,6 @@
 data_frame_report <- function(frm,
                               chunk_size = 1e2,
-                              to_remove = c()) {
+                              to_ignore = c(),
                               warn_if_missing=F) {
   #' Proper Noun Detection
   #'
@@ -11,14 +11,18 @@ data_frame_report <- function(frm,
   #' @param frm A data frame to check for proper nouns
   #' @param chunk_size [optional] The number of sentences to tag at a time.  The optimal value
   #'   has yet to be determined.
-  #' @param to_remove [optional] A character vector of column names to remove from the data frame
+  #' @param to_ignore [optional] A vector of column names to be ignored by the algorithm.  
+  #'   Intended to be used for variables that are giving strong false positives, such as
+  #'   IDs or ICD-10 codes.
+  #' @param warn_if_missing [optional] Raise a warning if the `to_ignore` columns are 
+  #'   not in the data frame.
   #'
-  #' @return A tibble containing:
+  #' @return A `pid_report` (inheriting from tibble) containing:
   #' \itemize{
   #'   \item `ID`: The location of the sentence in the data frame in the form `Col:<colname> Row:<rownumber>`.
   #'   \item `Token`: The detected proper noun.
   #'   \item `Sentence`: The sentence containing the proper noun.
-  #'   \item Repats: The number of times the sentence occurs in the data frame.
+  #'   \item `Repeats`: The number of times the sentence occurs in the data frame.
   #'   \item `Affected Columns`: The columns in the data frame where the sentence occurs.
   #' }
   #' If no proper nouns are detected, an empty data frame is returned.
@@ -38,7 +42,7 @@ data_frame_report <- function(frm,
   #' 
   
   frm_cols <- colnames(frm)
-  cant_remove <- setdiff(to_remove, frm_cols)
+  cant_remove <- setdiff(to_ignore, frm_cols)
   
   if (warn_if_missing & (length(cant_remove) > 0)) {
     warning(
@@ -48,7 +52,7 @@ data_frame_report <- function(frm,
     )
   }
   
-  tags <- data_frame_tagger(frm, chunk_size, to_remove)
+  tags <- data_frame_tagger(frm, chunk_size, to_ignore)
 
   report <- left_join(
     tags$`Proper Nouns`,
