### 2026-01-06 bdb1257
Style code (GHA)
