+++ tests/testthat/test-read_data.R
@@ -28,7 +28,6 @@ test_that("read_data() works with tsv files", {
 })
 
 
-
 test_that("read_data() works with xlsx files", {
   local({
     .file <- local_tempfile(fileext = ".xlsx")
