### 2026-01-06 20d0c08
Amend example to run.
converted dontrun block to try method,.
