+++ R/data_frame_report.R
@@ -28,10 +28,11 @@ data_frame_report <- function(frm,
   #' If no proper nouns are detected, an empty data frame is returned.
   #'
   #' @examples
-  #' \dontrun{
+  #' data(the_one_in_massapequa)
   #' example.data <- head(the_one_in_massapequa, 10)
-  #' data_frame_report(example.data, to_remove=c("scene", "utterance"))
-  #' }
+  #' try(
+  #'   data_frame_report(example.data, to_ignore=c("scene", "utterance"))
+  #' )
   #' @export
   #' @importFrom magrittr %>%
   #' @importFrom dplyr group_by group_modify left_join where all_of
