### 2026-01-06 566c58d
Add JOSSReviewRecommendations branch to workflow
