+++ .github/workflows/R-CMD-check.yaml
@@ -2,7 +2,7 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, master, redesign]
+    branches: [main, master, redesign, JOSSReviewRecommendations]
     paths: ["**.R", "DSCRIPTION", "NAMESPACE"]
   pull_request:
 
