### 2026-01-06 5c1126f
Improve encode error message
Split error message to be more informative
