### 2026-01-06 2bb075b
Style code (GHA)
