### 2026-01-07 #9728 @sebasquirarte
- [x] **Repository:** Is the source code for this software available at https://github.com/Stat-Cook/pid.pos?
- [x] **Contribution and authorship:** Has the submitting author ({{author-handle}}) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
