### 2026-01-09 Stat-Cook/pid.pos#6 @Stat-Cook
@agerada We've been pondering the way of including the udpipe model. Part of the aim of this package is to supply a tool for use by those with minimal technical skills (NHS BI/ research teams) so wanted to automate the inclusion of the model as much as possible.  What would you think if we removed the inclusion of the model at install from the main branch and had a secondary branch (e.g. ReadyOutOfTheBox) configured for automatic set up?
