### 2026-01-28 #9728 @sebasquirarte
I will open an issue in the GitHub repository shortly for comments and feedback related specifically to the code, but I wanted to first mention some general comments and suggestions regarding the repository and paper:

Minor grammar and formatting issues in the paper that should be addressed:

>
> Line 8 – “The extraction of proper nouns reduced […]” should be “The extraction of proper nouns reduces […]”
> Line 28 – “[...] framework(Straka et al. […]” is missing a space between 'framework' and the citation
> Line 58 – “The vignette … is included [...]” This specific vignette does not appear to be included
> Line 60 – There appears to be an unnecessary line break at the end of this line, between the words 'supply' and 'basic'
>

Although there is a brief “Getting Help” section in the repository’s README, I would suggest adding a CONTRIBUTING.md file and a section outlining contribution guidelines, code of conduct, and related information.

It is good that the paper includes some examples and that vignettes are available, however these are quite brief and somewhat superficial. They could be expanded and improved to better illustrate the package’s functionality and the use of its different functions.

If possible, submitting the package to CRAN would help validate its quality and usability, ensure alignment with best practices, and facilitate integration within the R ecosystem.

Finally, I suggest adding the missing DOI mentioned by the editorial bot.
