### 2026-01-28 #9728 @agerada
> Hi [@agerada](https://github.com/agerada) and [@sebasquirarte](https://github.com/sebasquirarte) how is the review process going on. Please let me know if you have any questions. Thanks

Hi @adithirgis. My initial review is available over here: https://github.com/Stat-Cook/pid.pos/issues/6

I am monitoring the open issue, and additionally will submit a review of the paper soon.
