+++ man/all_random_replacement.f.Rd
@@ -22,7 +22,7 @@ of the same input.
 \examples{
 
 replace_by <- all_random_replacement.f()
-auto_replace(raw_redaction_rules, replacement.f = replace_by) 
+auto_replace(raw_redaction_rules, replacement.f = replace_by)
 
 replace_by <- all_random_replacement.f(replacement_space = LETTERS[1:10], replacement_size = 20)
 auto_replace(raw_redaction_rules, replacement.f = replace_by)
