+++ R/browse.R
@@ -1,4 +1,4 @@
-browse_udpipe_repo <- function() {
+browse_udpipe_repo <- function(model = "english-ewt") {
   #' Open github link to the 'english-ewt-2.5' UD model.
   #'
   #' Intended for user download where `udpipe` fails to download automatically.
@@ -6,9 +6,13 @@ browse_udpipe_repo <- function() {
   #'
   #' @export
   #' @importFrom utils browseURL
-  udpipe_repo <- "https://github.com/jwijffels/udpipe.models.ud.2.5/blob/master/inst/udpipe-ud-2.5-191206/english-ewt-ud-2.5-191206.udpipe"
+  
+  url_root <- "https://github.com/jwijffels/udpipe.models.ud.2.5/blob/master/inst/udpipe-ud-2.5-191206"
+  url_path <- sprintf("%s-ud-2.5-191206.udpipe", model)
+  udpipe_repo <- paste(url_root, url_path, sep = "/")
   browseURL(udpipe_repo)
 }
+
 
 browse_model_location <- function() {
   #' Browse user to folder for UDPipe models.
