### 2026-02-03 39fa817
Functional re-design.
