+++ dev/redesign.R
@@ -1,0 +1,227 @@
+tagger_factory <- function(model = "english-ewt", model_dir = pid.pos_env$model_folder) {
+  #' 
+  #' Creates a tagging function using the specified UDPipe model.
+  #' 
+  #' @param model The UDPipe model to use for tagging.
+  #' @param model_dir The directory where the UDPipe model is stored.
+  #' @return A function that tags documents using the specified UDPipe model.
+  #' 
+  #' @importFrom udpipe udpipe
+  #' @importFrom dplyr mutate
+  #' @export
+  #' @example 
+  #' tagger <- tagger_factory("english-ewt")
+  #' tagged <- tagger(c("This is a test.", "Another sentence."), doc_id = c("doc1", "doc2"))
+  #' tagged
+  function(docs,
+           doc_id = NULL,
+           catch = T) {
+    if (is.null(doc_id)) {
+      .doc_id <- seq_along(docs)
+    }
+    
+    if (is.numeric(doc_id)) {
+      .doc_id <- paste("doc", doc_id, sep = "")
+    }
+    
+    utf8_docs <- utf8_encode(docs)
+    names(utf8_docs) <- doc_id
+    
+    if (catch) {
+      tagged <- tryCatch(
+        udpipe(utf8_docs, model, model_dir = model.dir),
+        error = function(e)
+          stop(
+            "UDPipe Model can't be loaded/ downloaded.
+                             Please run `browse_model_location()` to see if models are downloaded
+                             and if not present download via `browse_udpipe_repo()`."
+          )
+      )
+    } else {
+      tagged <- udpipe(utf8_docs, model, 
+                       model_dir = model_dir)
+    }
+    
+    mutate(tagged, `Token No` = as.numeric(.data$`token_id`))
+  }
+}
+
+
+tag_documents <- function(docs,
+                          doc_ids,
+                          tagger = tagger_factory(),
+                          chunk_size = 100) {
+
+  if (is.null(doc_ids)) {
+    doc_ids <- seq_along(docs)
+  }
+  
+  n <- length(docs)
+  
+  splits <- floor((1:n - 1) / chunk_size)
+  jobs <- split(docs, splits)
+  ids <- split(doc_ids, splits)
+  
+  tagged <- map2(jobs, ids, function(docs, ids)
+    tagger(docs, ids), .progress = T)
+  bind_rows(tagged)
+}
+
+
+tag_data_frame <- function(frm,
+                           tagger = "english-ewt",
+                           chunk_size = 1e2,
+                           to_ignore = c()) {
+  #' Tags a data frame with part of speech tags
+  #'
+  #' @param frm A data frame to tag
+  #' @param tagger Either a string naming a UDPipe model (see `tagger_factory` for the list of models)
+  #'   or a custom tagging function (see `tagger_factory` for details of what is required).
+  #' @param chunk_size The number of sentences to tag at a time
+  #' @param to_ignore A character vector of column names to remove from the data frame
+  #'
+  #' @importFrom dplyr group_by group_modify row_number ungroup
+  #' @importFrom glue glue
+  #' @importFrom progress progress_bar
+  #' @importFrom dplyr where all_of filter
+  #' @importFrom purrr simplify
+  
+  UseMethod("tag_data_frame", tagger)
+  
+}
+
+
+tag_data_frame.character <- function(frm,
+                                     tagger,
+                                     chunk_size = 1e2,
+                                     to_ignore = c()) {
+  tagger <- tagger_factory(tagger)
+  tag_data_frame(frm, tagger, chunk_size, to_ignore)
+}
+
+
+tag_data_frame.function <- function(frm,
+                                    tagger,
[…]
