### 2026-02-03 1504b8a
Fix typo in docstring
