### 2026-02-03 bff715b
Style code (GHA)
