+++ R/zzz.R
@@ -7,7 +7,7 @@
   )
   toset <- !(names(op.pid.pos) %in% names(op))
   if (any(toset)) options(op.pid.pos[toset])
-  
+
   pid.pos_env$deault_model_folder <- user_data_dir("pid.pos")
   enable_local_models()
 
