### 2026-02-03 e23a24f
Style code (GHA)
