+++ dev/redesign.R
@@ -1,4 +1,4 @@
-tagger_factory <- function(model = "english-ewt", model_dir = pid.pos_env$model_folder) {
+tagger_factory <- function(model = "english-ewt", model.dir = pid.pos_env$model_folder) {
   #' 
   #' Creates a tagging function using the specified UDPipe model.
   #' 
@@ -10,12 +10,18 @@ tagger_factory <- function(model = "english-ewt", model_dir = pid.pos_env$model_
   #' @importFrom dplyr mutate
   #' @export
   #' @examples
-  #' tagger <- tagger_factory("english-ewt")
-  #' tagged <- tagger(c("This is a test.", "Another sentence."), doc_id = c("doc1", "doc2"))
-  #' tagged
+  #' ewt_tagger <- tagger_factory("english-ewt")
+  #' ewt_tagger(c("This is a test.", "Another sentence."))
+  #' 
+  #' gum_tagger <- tagger_factory("english-gum")
+  #' gum_tagger(c("This is a test.", "Another sentence."))
+  #' 
+  #' lines_tagger <- tagger_factory("english-lines")
+  #' lines_tagger(c("This is a test.", "Another sentence."))
+  #' 
+  #' 
   function(docs,
-           doc_id = NULL,
-           catch = T) {
+           doc_id = NULL) {
     if (is.null(doc_id)) {
       .doc_id <- seq_along(docs)
     }
@@ -26,21 +32,17 @@ tagger_factory <- function(model = "english-ewt", model_dir = pid.pos_env$model_
     
     utf8_docs <- utf8_encode(docs)
     names(utf8_docs) <- doc_id
-    
-    if (catch) {
-      tagged <- tryCatch(
-        udpipe(utf8_docs, model, model_dir = model.dir),
-        error = function(e)
-          stop(
-            "UDPipe Model can't be loaded/ downloaded.
-                             Please run `browse_model_location()` to see if models are downloaded
-                             and if not present download via `browse_udpipe_repo()`."
-          )
-      )
-    } else {
-      tagged <- udpipe(utf8_docs, model, 
-                       model_dir = model_dir)
-    }
+
+    tagged <- tryCatch(
+      udpipe(utf8_docs, model, model_dir = model.dir),
+      error = function(e){
+        print(e$message)
+        stop(
+          "UDPipe Model can't be loaded/ downloaded.
+                           Please run `browse_model_location()` to see if models are downloaded
+                           and if not present download via `browse_udpipe_repo()`."
+        )
+    })
     
     mutate(tagged, `Token No` = as.numeric(.data$`token_id`))
   }
@@ -48,10 +50,21 @@ tagger_factory <- function(model = "english-ewt", model_dir = pid.pos_env$model_
 
 
 tag_documents <- function(docs,
-                          doc_ids,
+                          doc_ids=NULL,
                           tagger = tagger_factory(),
                           chunk_size = 100) {
-
+  #' @examples
+  #' example.text <- head(the_one_in_massapequa$text, 20)
+  #' 
+  #' ewt_tagger <- tagger_factory("english-ewt")
+  #' tag_documents(example.text, tagger=ewt_tagger)
+  #' 
+  #' gum_tagger <- tagger_factory("english-gum")
+  #' tag_documents(example.text, tagger=gum_tagger)
+  #' 
+  #' lines_tagger <- tagger_factory("english-lines")
+  #' tag_documents(example.text, tagger=lines_tagger)
+  #' 
   if (is.null(doc_ids)) {
     doc_ids <- seq_along(docs)
   }
@@ -63,7 +76,7 @@ tag_documents <- function(docs,
[…]
