+++ CONTRIBUTING.md
@@ -1,0 +1,30 @@
+# Contributing to pid.pos
+
+This outlines how to contribute to the pid.pos package.
+
+## General contributions
+
+Contributions to pid.pos are welcome, and intended to the scale of the change. 
+- Typos: feel free to correct these using the GitHub web-interface or by forking the repository. 
+  The project uses ROxygen so please make changes in the `.R` files *not* in the `.Rd` files. 
+- Bug reports: please file a minimal reproducible example on GitHub.
+- Feature requests: please open an issue on GitHub outlining the feature you would like to see, and ideally a use case for it.
+- Implemented features: if you are thinking of expanding the package with a new feature, please open an issue first to discuss the 
+  design principles and ensure it is compatible with the overall aims of the package.
+
+## Pull request process
+
+* Fork the repository on GitHub
+* Clone your forked repository to your local machine
+* Create a new branch for your changes (`git checkout -b my-new-feature`)
+* Make your changes and commit them (`git commit -am 'Add some feature'`)
+* Push your changes to your forked repository (`git push origin my-new-feature`)
+* Create a new Pull Request on GitHub
+* Ensure that all checks pass (e.g., R CMD check, tests, etc.)
+* Address any review comments from the maintainers
+* Once approved, your changes will be merged into the main repository
+* Thank you for contributing to pid.pos!
+
+## Code of Conduct
+This project follows the [Contributor Covenant Code of Conduct](CODE_OF_CONDUCT.md). 
+By participating, you agree to abide by its terms.
