### 2026-02-13 45dfd7a
Add model download test
