### 2026-02-13 b804503
Implement tag_documents
