### 2026-02-13 3ea9fab
Improve error propogation
