### 2026-02-13 bd69ccd
Add mockery calls
