+++ tests/testthat/test-udpipe_factory.R
@@ -1,13 +1,13 @@
 local({
   tagger <- udpipe_factory("english-ewt")
-  stub(tagger, "udpipe::udpipe", fake_udpipe_success)
+  mockery::stub(tagger, "udpipe::udpipe", fake_udpipe_success)
   
   test_that("tagger processes documents correctly", {
     result <- tagger("Test sentence")
     
     expect_s3_class(result, "tbl_df")
     expect_setequal(names(result),
-                    c("ID", "token_id", "Token", "Sentence", "TokenNo"))
+                    c("ID", "token_id", "Token", "Sentence", "TokenNo", "upos"))
     expect_equal(result$Token, "Test")
     expect_equal(result$ID, "doc1")
   })
@@ -25,7 +25,7 @@ local({
 test_that("tagger errors cleanly when UDPipe fails", {
   # Inject the failing fake
   tagger <- udpipe_factory(model = "english-ewt")
-  stub(tagger, "udpipe::udpipe", fake_udpipe_failure)
+  mockery::stub(tagger, "udpipe::udpipe", fake_udpipe_failure)
   
   # Expect a controlled error with your custom message
   expect_error(tagger("This is a test."))
@@ -34,7 +34,7 @@ test_that("tagger errors cleanly when UDPipe fails", {
 test_that("tagger can download model", {
   # Inject the failing fake
   tagger <- udpipe_factory(model = "english-ewt")
-  # Expect a controlled error with your custom message
+  
   result <- tagger("This is a test.")
   expect_equal(nrow(result), 5)
 })
