### 2026-02-16 b8ccdfd
Replace functional argument.
