### 2026-02-16 40fa539
Implement `tag_data_frame`
