### 2026-02-16 ab0a905
Implement and isolate 'File not found' error
