### 2026-02-16 0402293
Implement `type_error`
