### 2026-02-16 08b710d
Implement `make_unique_cols`
