+++ R/validation_error.R
@@ -1,0 +1,2 @@
+
+validation_error <- new_error_type("validation_error")
