+++ R/type_error.R
@@ -1,7 +1,6 @@
 #' A custom abort function 
 #' 
-#' @param msg The error message to display
-#' @param inherited_class A character vector of additional classes to inherit from (optional)
+#' @param message The error message to display
 #' @param ... Additional arguments to pass to `abort()`
 #' @param call The call environment to use for the error (defaults to the caller's environment)
 #' 
@@ -13,21 +12,10 @@
 #' @examples 
 #' #' # Example of using type_error
 #' f <- function(){
-#'   type_error("This is a type error", call=rlang::caller_env())
+#'   type_error("This is a type error")
 #' }
 #' 
 #' # Trigger the error safely without stopping R
 #' result <- try(f(), silent = TRUE)
 #' result
-#' 
-type_error <- function(msg, inherited_class = NULL, ..., call = caller_env()) {
-
-  cls <- c("pid_pos_type_error", "type_error", "pid_pos_error", inherited_class)
-  
-  abort(
-    message = msg,
-    class = cls,
-    call = call,
-    !!!list(...)
-  )
-}
+type_error <- new_error_type("type_error")
