### 2026-02-16 111d1b1
Implement 'filter_to_proper_nouns'
