### 2026-02-16 4dd622c
Implement `pid_pod` function
