+++ NAMESPACE
@@ -15,13 +15,20 @@ export(browse_model_location)
 export(browse_udpipe_repo)
 export(data_frame_report)
 export(efficient_redaction)
+export(enable_local_models)
+export(enable_package_models)
+export(filter_to_proper_nouns)
 export(frame_replacement)
 export(hashing_replacement.f)
 export(merge_redactions)
+export(pid_pos)
 export(prepare_redactions)
 export(random_replacement.f)
 export(report_on_folder)
 export(report_to_redaction_rules)
+export(set_udpipe_version)
+export(tag_documents)
+export(udpipe_factory)
 importFrom(R6,R6Class)
 importFrom(dplyr,all_of)
 importFrom(dplyr,arrange)
@@ -51,6 +58,8 @@ importFrom(purrr,reduce)
 importFrom(purrr,simplify)
 importFrom(rappdirs,user_data_dir)
 importFrom(readxl,read_excel)
+importFrom(rlang,abort)
+importFrom(rlang,caller_env)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_extract_all)
 importFrom(stringr,str_length)
@@ -60,6 +69,7 @@ importFrom(stringr,str_replace)
 importFrom(stringr,str_replace_all)
 importFrom(stringr,str_sub)
 importFrom(tibble,as_tibble)
+importFrom(tidyr,pivot_longer)
 importFrom(udpipe,udpipe)
 importFrom(udpipe,udpipe_load_model)
 importFrom(utf8,utf8_encode)
