+++ R/pid_pos.R
@@ -8,7 +8,7 @@
 #' @param tagger [optional] Either a string naming a UDPipe model (see ... for the list of models)
 #'   or a custom tagging function (see ... for details of what is required).
 #' @param filter_func [optional] A function that takes a tagged data frame and returns
-#'   
+#'
 #' @param chunk_size [optional] The number of sentences to tag at a time.  The optimal value
 #'   has yet to be determined.
 #' @param to_ignore [optional] A vector of column names to be ignored by the algorithm.
@@ -31,21 +31,23 @@
 #' data(the_one_in_massapequa)
 #' example.data <- head(the_one_in_massapequa, 20)
 #' try(
-#'   pid_pos(example.data, to_ignore=c("scene", "utterance"))
+#'   pid_pos(example.data, to_ignore = c("scene", "utterance"))
 #' )
-#' 
-#' pid_pos(example.data, to_ignore=c("scene", "utterance"), tagger="english-gum")
-#' 
+#'
+#' pid_pos(example.data, to_ignore = c("scene", "utterance"), tagger = "english-gum")
+#'
 #' tag_ewt <- udpipe_factory("english-ewt")
-#' pid_pos(example.data, to_ignore=c("scene", "utterance"), tagger=tag_ewt)
-#' 
-#' filter_to_long_proper_nouns <- function(frm){
+#' pid_pos(example.data, to_ignore = c("scene", "utterance"), tagger = tag_ewt)
+#'
+#' filter_to_long_proper_nouns <- function(frm) {
 #'   filter_to_proper_nouns(frm) |>
 #'     filter(nchar(Token) > 1)
 #' }
-#' pid_pos(example.data, to_ignore=c("scene", "utterance"), 
-#'   tagger=tag_ewt, filter=filter_to_long_proper_nouns)
-#' 
+#' pid_pos(example.data,
+#'   to_ignore = c("scene", "utterance"),
+#'   tagger = tag_ewt, filter = filter_to_long_proper_nouns
+#' )
+#'
 #' @export
 #' @importFrom magrittr %>%
 #' @importFrom dplyr group_by group_modify left_join where all_of
@@ -60,43 +62,42 @@ pid_pos <- function(frm,
                     chunk_size = 1e2,
                     to_ignore = c(),
                     warn_if_missing = FALSE) {
-  
   if (!is.data.frame(frm)) {
     stop("`frm` must be a data frame.", call. = FALSE)
   }
-  
+
   if (!is.function(filter_func)) {
     stop("`filter_func` must be a function.", call. = FALSE)
   }
-  
+
   frm_cols <- colnames(frm)
   cant_remove <- setdiff(to_ignore, frm_cols)
-  
+
   if (warn_if_missing && (length(cant_remove) > 0)) {
     warning(
       glue(
         "The following columns to remove were not found in the data frame: {paste(cant_remove, collapse=', ')}",
         call. = FALSE
       )
-      )
+    )
   }
-  
+
   tagged <- tag_data_frame(frm, tagger, chunk_size, to_ignore)
-  
+
   if (is.null(tagged$AllTags) || is.null(tagged$Documents)) {
     return(structure(
       tibble::tibble(),
       class = c("pid_report", "tbl_df", "tbl", "data.frame")
     ))
   }
-  
+
   filtered_tags <- filter_func(tagged$AllTags)
-  
+
   report <- left_join(filtered_tags, tagged$Documents, by = "ID") %>%
     as_tibble() |>
     arrange(PK) |>
     select(-PK)
-  
+
   structure(
     report,
     class = c("pid_report", class(report))
