### 2026-02-17 3020d01
deprecate `date_frame_report`
