+++ R/frame_replacement.R
@@ -79,7 +79,7 @@ frame_replacement <- function(frm, rules.frm) {
   #' @examples
   #' \dontrun{
   #' example.data <- head(the_one_in_massapequa)
-  #' report <- data_frame_report(example.data)
+  #' report <- pid_pos(example.data)
   #' redactions.raw <- report_to_redaction_rules(report)
   #'
   #' replace_by <- random_replacement.f()
