+++ R/efficient_redact_factory.R
@@ -84,7 +84,7 @@ efficient_redaction <- function(frm, redact, n = NULL, .progress = T) {
   #' @examples
   #' \dontrun{
   #' example.data <- head(the_one_in_massapequa)
-  #' report <- data_frame_report(example.data, to_remove="speaker")
+  #' report <- pid_pos(example.data, to_remove="speaker")
   #' redactions.raw <- report_to_redaction_rules(report)
   #'
   #' replace_by <- random_replacement.f()
