+++ R/merge_replacements.R
@@ -1,4 +1,4 @@
-vector_merge_redactions <- function(vec, cached_redactions, preprocess = utf8_encode) {
+vector_merge_redactions <- function(vec, cached_redactions, preprocess = utf8::utf8_encode) {
   #' @importFrom dplyr left_join join_by
   frm <- data.frame(Old = preprocess(vec)) |>
     left_join(cached_redactions, by = dplyr::join_by(Old == If)) |>
@@ -8,7 +8,7 @@ vector_merge_redactions <- function(vec, cached_redactions, preprocess = utf8_en
 }
 
 
-merge_redactions <- function(frm, cached_redactions, preprocess = utf8_encode) {
+merge_redactions <- function(frm, cached_redactions, preprocess = utf8::utf8_encode) {
   #' Remove PID from a data frame via a merge/
   #'
   #' @param frm The data frame to be redacted
