+++ R/report_to_redaction_rules.R
@@ -1,0 +1,66 @@
+#' Initialize redaction rules
+#'
+#' Convert the output of `pid_pos` to a csv file for editing.
+#' It is intended the user changes the `To` column to define desired
+#' redactions.
+#'
+#' @param report The `data.frame` generated by `pid_pos`
+#' @param path [optional] File path for write.
+#' @param include_context [optional] Binary-flag. If True, adds a `Context` column giving a subset
+#' of the sentence being redacted around the token.
+#'
+#' @return A tibble with columns:
+#' \itemize{
+#' \item `If`: The condition (sentence) under which the replacement is to be made.
+#' \item `From`: The original token.
+#' \item `To`: The replacement token.
+#' }
+#'
+#' @examples
+#' \dontrun{
+#' example.data <- head(the_one_in_massapequa)
+#' report <- pid_pos(example.data, to_remove = "speaker")
+#' report_to_redaction_rules(report)
+#' }
+#'
+#' @export
+report_to_redaction_rules <- function(report, path = NULL,
+                                      include_context = FALSE) {
+
+  required_cols <- c("Sentence", "Token")
+  missing <- setdiff(required_cols, names(report))
+  if (length(missing) > 0) {
+    validation_error("report is missing columns: ", paste(missing, collapse = ", "))
+  }
+  
+  if (!inherits(report, "data.frame")) {
+    type_error("`report` must be a data.frame or tibble")
+  }
+  
+  if (!is.null(path) && !is.character(path)) {
+    type_error("`path` must be a string or NULL")
+  }
+  
+  
+  .frm <- report |>
+    mutate(
+      If = Sentence,
+      From = Token,
+      To = Token,
+      .keep = "none"
+    )
+  
+  if (include_context) {
+    .frm <- .frm |>
+      mutate(
+        Context = map2(If, From, get_context)
+      )
+  }
+  
+  if (is.null(path)) {
+    return(.frm)
+  }
+  
+  readr::write_csv(.frm, path)
+  .frm
+}
