+++ R/report_as_frame.R
@@ -1,68 +1,0 @@
-report_to_redaction_rules <- function(report, path = NULL,
-                                      add_context = F) {
-  #' Initialize redaction rules
-  #'
-  #' Convert the output of `data_frame_report` to a csv file for editing.
-  #' It is intended the user changes the `To` column to define desired
-  #' redactions.
-  #'
-  #' @param report The `data.frame` generated by `data_frame_report`
-  #' @param path [optional] File path for write.
-  #' @param add_context [optional] Binary-flag. If True, adds a `Context` column giving a subset
-  #' of the sentence being redacted around the token.
-  #'
-  #' @return A tibble with columns:
-  #' \itemize{
-  #' \item `If`: The condition (sentence) under which the replacement is to be made.
-  #' \item `From`: The original token.
-  #' \item `To`: The replacement token.
-  #' }
-  #'
-  #' @examples
-  #' \dontrun{
-  #' example.data <- head(the_one_in_massapequa)
-  #' report <- data_frame_report(example.data, to_remove = "speaker")
-  #' report_to_redaction_rules(report)
-  #' }
-  #'
-  #' @export
-  .frm <- report |>
-    mutate(
-      # If = Document,
-      If = Sentence,
-      From = Token,
-      To = Token,
-      .keep = "none"
-    )
-
-  if (add_context) {
-    .frm <- .frm |>
-      mutate(
-        Context = get_context(If, From)
-      )
-  }
-
-  if (is.null(path)) {
-    return(.frm)
-  }
-
-  write.csv(.frm, path)
-  .frm
-}
-
-
-#
-# if_modify <- function(frm, group) {
-#   .if <- group$If
-#
-#   .left <- sprintf("stringr::str_detect(.x, '%s')", .if)
-#   # .left <- sprintf(".x == '%s'", .if)
-#   .right <- paste(c(".x", frm$Replace), collapse = " |> ")
-#
-#
-#   sprintf("%s ~ %s", .left, .right)
-# }
-#
-# escape_quote_mark <- function(vec) {
-#   str_replace_all(vec, "'", "\\\\'")
-# }
