### 2026-02-17 f1142cb
Implement as single function for use with map2
