### 2026-02-17 7708525
Deprecate `data_frame_report`
