### 2026-02-17 e6b7297
Deprecate `pos_tag`
