+++ NAMESPACE
@@ -13,7 +13,6 @@ export(all_random_replacement.f)
 export(auto_replace)
 export(browse_model_location)
 export(browse_udpipe_repo)
-export(data_frame_report)
 export(efficient_redaction)
 export(enable_local_models)
 export(enable_package_models)
@@ -70,9 +69,7 @@ importFrom(stringr,str_replace_all)
 importFrom(stringr,str_sub)
 importFrom(tibble,as_tibble)
 importFrom(tidyr,pivot_longer)
-importFrom(udpipe,udpipe)
 importFrom(udpipe,udpipe_load_model)
-importFrom(utf8,utf8_encode)
 importFrom(utils,browseURL)
 importFrom(utils,head)
 importFrom(utils,read.csv)
