### 2026-02-17 be3044e
Implement `custom_tagger` wrapper
