### 2026-02-17 0570501
remove vestigial file
