### 2026-02-17 ecb954d
remove deprecated tests
