+++ tests/testthat/test-summary.R
@@ -1,5 +1,5 @@
 test_that("summary works", {
-  report <- data_frame_report(the_one_in_massapequa)
+  report <- pid_pos(the_one_in_massapequa)
   .summary <- summary(report)
 
   expect_equal(nrow(.summary), 2)
