+++ tests/testthat/test-make_replacements.R
@@ -1,5 +1,5 @@
 test_that("make and merge replacements workflow", {
-  report <- data_frame_report(the_one_in_massapequa)
+  report <- pid_pos(the_one_in_massapequa)
 
   .rules <- report_to_redaction_rules(
     report
