### 2026-02-17 60e7e1d
Implement sentence levels tests
