+++ tests/testthat/test-get_context.R
@@ -1,22 +1,22 @@
 test_that("get_context works", {
-  .subset <- the_one_in_massapequa[101:110, ]
-
-  text.responses <- data_frame_report(.subset) |>
-    filter(`Affected Columns` == "`text`")
-
-  set_context_window(5)
-  context <- get_context(text.responses$Sentence, text.responses$Token)
-
-  expect_equal(nrow(text.responses), length(context))
-
-  expect_true(
-    all(str_detect(context, text.responses$Token))
+  expect_match(
+    get_context("This is a sentence to be tested", "sentence"),
+    "This is a sentence to be tested"
   )
-
-  set_context_window(0)
-  context.0 <- get_context(text.responses$Sentence, text.responses$Token)
-
-  expect_true(
-    all(context.0 != context)
+  
+  expect_equal(
+    get_context("This is a sentence to be tested", "sentence", 2),
+    "...a sentence t..."
   )
+  
+  set_context_window(2)
+  
+  expect_equal(
+    get_context("This is a sentence to be tested", "sentence"),
+    "...a sentence t..."
+  )
+  
+  set_context_window(25)
+  
 })
+  
