+++ tests/testthat/test-get_context.R
@@ -3,20 +3,18 @@ test_that("get_context works", {
     get_context("This is a sentence to be tested", "sentence"),
     "This is a sentence to be tested"
   )
-  
+
   expect_equal(
     get_context("This is a sentence to be tested", "sentence", 2),
     "...a sentence t..."
   )
-  
+
   set_context_window(2)
-  
+
   expect_equal(
     get_context("This is a sentence to be tested", "sentence"),
     "...a sentence t..."
   )
-  
+
   set_context_window(25)
-  
 })
-  
