+++ tests/testthat/test-custom_tagger.R
@@ -4,8 +4,11 @@ dummy_pos <- function(doc) {
 }
 
 dummy_empty <- function(doc) {
-  if (nchar(doc) == 0) tibble::tibble(word = character(), pos = character()) 
-  else tibble::tibble(word = unlist(strsplit(doc, " ")), pos = "NN")
+  if (nchar(doc) == 0) {
+    tibble::tibble(word = character(), pos = character())
+  } else {
+    tibble::tibble(word = unlist(strsplit(doc, " ")), pos = "NN")
+  }
 }
 
 # Wrap the dummy POS functions
@@ -18,11 +21,11 @@ docs_special <- c("Hello, world!", "R&D is fun")
 
 test_that("basic functionality works", {
   res <- tagger(docs)
-  
+
   expect_equal(nrow(res), sum(sapply(strsplit(docs, " "), length)))
-  
+
   expect_true(all(c("word", "pos", "doc_id") %in% colnames(res)))
-  
+
   expect_true(all(res$doc_id %in% 1:2))
 })
 
@@ -33,7 +36,7 @@ test_that("custom doc_ids are used", {
 
 test_that("POS function returning empty data frame is handled", {
   res_edge <- tagger_empty(c("Test", ""))
-  expect_equal(nrow(res_edge), 1)  # only one word from non-empty doc
+  expect_equal(nrow(res_edge), 1) # only one word from non-empty doc
   expect_true(all(c("word", "pos", "doc_id") %in% colnames(res_edge)))
 })
 
@@ -50,11 +53,9 @@ test_that("results are consistent across multiple runs", {
 })
 
 test_that("type check warnings", {
-  
   expect_error(custom_tagger(12))
   expect_error(custom_tagger("12"))
-  
+
   expect_error(custom_tagger(function() 12), "at least one argument")
   expect_error(custom_tagger(function(a, b) 12), "can't have more than one")
 })
-
