### 2026-02-17 9f83cad
implement `pid_pos` function
