+++ vignettes/PaperVignette.Rmd
@@ -88,7 +88,7 @@ First, the `data_frame_report` function converts a typical R data frame into a n
 * `Affected Columns` - the columns in the original data frame which contain the sentence
 
 ``` {r, message=F, warning = F}
-report <- data_frame_report(the_one_in_massapequa)
+report <- pid_pos(the_one_in_massapequa)
 report
 ```
 
