+++ vignettes/AutoReplacement.Rmd
@@ -41,7 +41,7 @@ The basic workflow is hence:
 ```{r, results="hide", message=F}
 library(pid.pos)
 
-report <- data_frame_report(the_one_in_massapequa)
+report <- pid_pos(the_one_in_massapequa)
 replacement_rules <- report_to_redaction_rules(report)
 
 replacement_rules
