### 2026-02-17 554e1eb
Add defensive checks
