### 2026-02-17 f9b6111
Update for local_models, set_updipe_version
