### 2026-02-18 Stat-Cook/pid.pos#6 @agerada
On the issue of firewalls, I completely understand. Let's say that the user decides to provide the model themselves (for example by downloading on a different machine). Can they then use that specific model with `pid.pos`, either as a local path or as a loaded object?
