### 2026-02-19 63e0f8c
Ensure  feature space is added.
