### 2026-02-19 910856a
Make internal to allow examples
