+++ NAMESPACE
@@ -13,6 +13,7 @@ export(all_random_replacement.f)
 export(auto_replace)
 export(browse_model_location)
 export(browse_udpipe_repo)
+export(custom_tagger)
 export(efficient_redaction)
 export(enable_local_models)
 export(enable_package_models)
@@ -26,6 +27,7 @@ export(random_replacement.f)
 export(report_on_folder)
 export(report_to_redaction_rules)
 export(set_udpipe_version)
+export(tag_data_frame)
 export(tag_documents)
 export(udpipe_factory)
 importFrom(R6,R6Class)
