### 2026-02-19 6bde012
Implement CustomFunction vignette
