### 2026-02-19 cc0598a
Style code (GHA)
