### 2026-02-20 bb5e1fd
Move model to localdir to enable vignette build.
