### 2026-02-20 e9d046a
Fix static coding querries
