### 2026-02-20 ba9bfca
Match tests to change in implementation.
