+++ NAMESPACE
@@ -31,6 +31,7 @@ export(tag_data_frame)
 export(tag_documents)
 export(udpipe_factory)
 importFrom(R6,R6Class)
+importFrom(dplyr,across)
 importFrom(dplyr,all_of)
 importFrom(dplyr,arrange)
 importFrom(dplyr,bind_rows)
@@ -59,6 +60,8 @@ importFrom(purrr,reduce)
 importFrom(purrr,simplify)
 importFrom(rappdirs,user_data_dir)
 importFrom(readxl,read_excel)
+importFrom(rlang,":=")
+importFrom(rlang,.data)
 importFrom(rlang,abort)
 importFrom(rlang,caller_env)
 importFrom(stringr,str_detect)
