+++ R/browse.R
@@ -8,11 +8,11 @@ browse_udpipe_repo <- function(model = "english-ewt") {
   #' @export
   #' @importFrom utils browseURL
   #'
-  
+
   .date <- pid.pos_env$repo_dates[[udpipe_repo]]
   .version <- pid.pos_env$udpipe_repo
   .version.number <- stringr::str_extract(.version, "\\d.\\d$")
-  
+
   url_root <- sprintf(
     "https://github.com/%s/blob/master/inst/udpipe-ud-%s-%s",
     .version,
@@ -20,9 +20,9 @@ browse_udpipe_repo <- function(model = "english-ewt") {
     .date
   )
   url_path <- sprintf("%s-ud-%s-%s.udpipe", model, .version.number, .date)
-  udpipe_repo <-       paste(url_root, url_path, sep = "/")
-  
-  
+  udpipe_repo <- paste(url_root, url_path, sep = "/")
+
+
   browseURL(udpipe_repo)
 }
 
@@ -33,6 +33,6 @@ browse_model_location <- function() {
   #' Intended for usage in `udpipe` fails to download automatically.
   #'
   #' @export
-  
+
   browseURL(pid.pos_env$model_folder)
 }
