+++ DESCRIPTION
@@ -45,6 +45,7 @@ Suggests:
     openxlsx,
     testthat (>= 3.0.0),
     mockery,
+    reticulate,
     covr,
     withr
 Config/testthat/edition: 3
