### 2026-02-20 9202555
Add workflow_dispatch trigger to R-CMD-check
