+++ man/enable_package_models.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{enable_package_models}
+\alias{enable_package_models}
+\title{Set the model folder to the package data folder.}
+\usage{
+enable_package_models()
+}
+\value{
+The path to the model folder.
+}
+\description{
+Intended if you want to share udpipe models between different R projects.
+}
+\examples{
+enable_package_models()
+}
