+++ man/efficient_redaction.Rd
@@ -30,7 +30,7 @@ This is useful for large data frames where the redaction function may be slow.
 \examples{
 \dontrun{
 example.data <- head(the_one_in_massapequa)
-report <- data_frame_report(example.data, to_remove="speaker")
+report <- pid_pos(example.data, to_remove="speaker")
 redactions.raw <- report_to_redaction_rules(report)
 
 replace_by <- random_replacement.f()
