+++ man/data_frame_tagger.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/data_frame_tagger.R
-\name{data_frame_tagger}
-\alias{data_frame_tagger}
-\title{Tags a data frame with part of speech tags}
-\usage{
-data_frame_tagger(frm, chunk_size = 100, to_remove = c())
-}
-\arguments{
-\item{frm}{A data frame to tag}
-
-\item{chunk_size}{The number of sentences to tag at a time}
-
-\item{to_remove}{A character vector of column names to remove from the data frame}
-}
-\description{
-Tags a data frame with part of speech tags
-}
