+++ man/prepare_redactions.Rd
@@ -20,7 +20,7 @@ to a function that can be applied to a data frame.
 \examples{
 \dontrun{
 example.data <- head(the_one_in_massapequa)
-report <- data_frame_report(example.data, to_remove="speaker")
+report <- pid_pos(example.data, to_remove="speaker")
 redactions.raw <- report_to_redaction_rules(report)
 
 replace_by <- random_replacement.f()
