+++ man/report_to_redaction_rules.Rd
@@ -1,17 +1,17 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/report_as_frame.R
+% Please edit documentation in R/report_to_redaction_rules.R
 \name{report_to_redaction_rules}
 \alias{report_to_redaction_rules}
 \title{Initialize redaction rules}
 \usage{
-report_to_redaction_rules(report, path = NULL, add_context = F)
+report_to_redaction_rules(report, path = NULL, include_context = FALSE)
 }
 \arguments{
-\item{report}{The `data.frame` generated by `data_frame_report`}
+\item{report}{The `data.frame` generated by `pid_pos`}
 
 \item{path}{[optional] File path for write.}
 
-\item{add_context}{[optional] Binary-flag. If True, adds a `Context` column giving a subset
+\item{include_context}{[optional] Binary-flag. If True, adds a `Context` column giving a subset
 of the sentence being redacted around the token.}
 }
 \value{
@@ -23,14 +23,14 @@ A tibble with columns:
 }
 }
 \description{
-Convert the output of `data_frame_report` to a csv file for editing.
+Convert the output of `pid_pos` to a csv file for editing.
 It is intended the user changes the `To` column to define desired
 redactions.
 }
 \examples{
 \dontrun{
 example.data <- head(the_one_in_massapequa)
-report <- data_frame_report(example.data, to_remove = "speaker")
+report <- pid_pos(example.data, to_remove = "speaker")
 report_to_redaction_rules(report)
 }
 
