+++ man/merge_redactions.Rd
@@ -4,7 +4,7 @@
 \alias{merge_redactions}
 \title{Remove PID from a data frame via a merge/}
 \usage{
-merge_redactions(frm, cached_redactions, preprocess = utf8_encode)
+merge_redactions(frm, cached_redactions, preprocess = utf8::utf8_encode)
 }
 \arguments{
 \item{frm}{The data frame to be redacted}
