+++ man/udpipe_factory.Rd
@@ -1,0 +1,49 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/udpipe_factory.R
+\name{udpipe_factory}
+\alias{udpipe_factory}
+\title{Create a UDPipe tagging function}
+\usage{
+udpipe_factory(
+  model = "english-ewt",
+  model_dir = pid.pos_env$model_folder,
+  udpipe_repo = pid.pos_env$udpipe_repo
+)
+}
+\arguments{
+\item{model}{Character. The name of the UDPipe model to use. Defaults to `"english-ewt"`.}
+
+\item{model_dir}{Character. Directory where UDPipe models are stored. Defaults to `pid.pos_env$model_folder`.}
+
+\item{udpipe_repo}{Character. URL or path of the UDPipe model repository. Defaults to `pid.pos_env$udpipe_version`.}
+}
+\value{
+A function that takes a character vector of documents and returns a `tibble`
+with columns:
+\describe{
+  \item{ID}{Document identifier}
+  \item{Token}{Individual token text}
+  \item{Sentence}{Sentence containing the token}
+  \item{upos}{The universal parts of speech tag of the token. See https://universaldependencies.org/format.html}
+}
+and all columns returned by the `udpipe::udpipe()` function for each token.
+}
+\description{
+Returns a function that tags text documents using a specified UDPipe model.
+The returned function accepts a character vector of documents and returns
+a tibble with tokens, sentences, and token metadata.
+}
+\examples{
+# Create a tagger for the English EWT model
+ewt_tagger <- udpipe_factory("english-ewt")
+docs <- c("This is a test.", "Another sentence.")
+ewt_tagger(docs)
+
+# Create a tagger for the English GUM model
+gum_tagger <- udpipe_factory("english-gum")
+gum_tagger(docs)
+
+# Create a tagger for the English LINES model
+lines_tagger <- udpipe_factory("english-lines")
+lines_tagger(docs)
+}
