+++ man/set_udpipe_version.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{set_udpipe_version}
+\alias{set_udpipe_version}
+\title{Set the udpipe model repository version.}
+\usage{
+set_udpipe_version(version = c("2.5", "2.4", "2.3"))
+}
+\arguments{
+\item{version}{Character. The udpipe model version to use. One of "2.5", "2.4", or "2.3".}
+}
+\value{
+Character. The udpipe model repository.
+}
+\description{
+Set the udpipe model repository version.
+}
+\examples{
+set_udpipe_version("2.4")
+}
