+++ man/file_not_found_error.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/file_not_found_error.R
+\name{file_not_found_error}
+\alias{file_not_found_error}
+\title{A custom abort function}
+\usage{
+file_not_found_error(message, ..., call = caller_env())
+}
+\arguments{
+\item{message}{The error message to display}
+
+\item{...}{Additional arguments to pass to `abort()`}
+
+\item{call}{The call environment to use for the error (defaults to the caller's environment)}
+}
+\value{
+An error object with the specified message and classes
+}
+\description{
+A custom abort function
+}
+\examples{
+#' # Example of using type_error
+f <- function() {
+  file_not_found_error("This is a type error", call = rlang::caller_env())
+}
+
+# Trigger the error safely without stopping R
+result <- try(f(), silent = TRUE)
+result
+
+}
+\keyword{internal}
