+++ man/filter_to_proper_nouns.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/filter_to_proper_nouns.R
+\name{filter_to_proper_nouns}
+\alias{filter_to_proper_nouns}
+\title{Filter a tagged data frame to proper nouns}
+\usage{
+filter_to_proper_nouns(tag_frm)
+}
+\arguments{
+\item{tag_frm}{A data frame containing at least the columns
+`upos`, `ID`, `Token`, and `Sentence`.}
+}
+\value{
+A tibble containing only rows where `upos == "PROPN"`,
+  with columns `ID`, `Token`, and `Sentence`.
+}
+\description{
+Filter a tagged data frame to proper nouns
+}
+\examples{
+example.data <- head(the_one_in_massapequa, 20)
+tagged <- tag_data_frame(example.data, tagger = "english-ewt")
+filter_to_proper_nouns(tagged$`AllTags`)
+
+}

[…]
