+++ tests/testthat/test-utilities.R
@@ -8,53 +8,67 @@ test_that("set_model_folder", {
 # Assume enable_local_models() and set_model_folder() are loaded
 
 test_that("creates and returns subfolder by default", {
-  withr::with_tempdir({
-    folder <- enable_local_models()
-
-    # Check folder was created
-    expect_true(dir.exists(file.path(getwd(), "pid_pos_models")))
-
-    # Check returned path
-    expect_equal(folder, file.path(getwd(), "pid_pos_models"))
-  })
+  withr::local_dir(withr::local_tempdir())
+  folder <- enable_local_models()
+  
+  # Check folder was created
+  expect_true(dir.exists(file.path(getwd(), "pid_pos_models")))
+  
+  # Check returned path
+  expect_equal(folder, file.path(getwd(), "pid_pos_models"))
+  
 })
 
 test_that("uses current working directory if sub_folder = FALSE", {
-  withr::with_tempdir({
-    folder <- enable_local_models(sub_folder = FALSE)
+  # Create a temporary working directory
+  withr::local_dir(withr::local_tempdir())
+  
+  folder <- enable_local_models(sub_folder = FALSE)
+  
+  # Should now point to the temp working directory
+  expect_equal(folder, getwd())
+  
+  # No pid_pos_models folder should exist
+  expect_false(dir.exists(file.path(getwd(), "pid_pos_models")))
+})
 
-    # Should be current working directory
-    expect_equal(folder, getwd())
-
-    # No pid_pos_models folder should be created
-    expect_false(dir.exists(file.path(getwd(), "pid_pos_models")))
-  })
+test_that("uses current working directory if sub_folder = FALSE", {
+  withr::local_dir(withr::local_tempdir())
+  folder <- enable_local_models(sub_folder = FALSE)
+  
+  # Should be current working directory
+  expect_equal(folder, getwd())
+  
+  # No pid_pos_models folder should be created
+  expect_false(dir.exists(file.path(getwd(), "pid_pos_models")))
+  
 })
 
 test_that("existing folder does not error", {
-  withr::with_tempdir({
-    path <- file.path(getwd(), "pid_pos_models")
-    dir.create(path)
-
-    # Should not error if folder exists
-    folder <- enable_local_models()
-    expect_equal(folder, path)
-  })
+  withr::local_dir(withr::local_tempdir())
+  
+  path <- file.path(getwd(), "pid_pos_models")
+  dir.create(path)
+  
+  # Should not error if folder exists
+  folder <- enable_local_models()
+  expect_equal(folder, path)
+  
 })
 
 test_that("returns folder path invisibly", {
-  withr::with_tempdir({
-    folder <- enable_local_models()
-
-    # Check return value is correct
-    expect_equal(folder, file.path(getwd(), "pid_pos_models"))
-  })
+  withr::local_dir(withr::local_tempdir())
+  folder <- enable_local_models()
+  
+  # Check return value is correct
+  expect_equal(folder, file.path(getwd(), "pid_pos_models"))
+  
 })
 
 test_that("integration with set_model_folder sets correct path", {
   withr::with_tempdir({
     folder <- enable_local_models()
-
+    
     # Assuming get_model_folder() returns the currently set folder
     if (exists("get_model_folder")) {
       expect_equal(get_model_folder(), folder)
@@ -66,20 +80,22 @@ test_that("integration with set_model_folder sets correct path", {
 })
 
 test_that("recursive folder creation works for nested subfolders", {
-  withr::with_tempdir({
[…]
