### 2026-02-20 48c40fd
Avoid creating folder in package on install.
