+++ R/utilities.R
@@ -16,20 +16,26 @@ enable_local_models <- function(sub_folder = TRUE) {
   #' @export
   #'
   #' @examples
-  #' enable_local_models()
-  #' enable_local_models(sub_folder=FALSE)
-
+  #' \dontrun{
+  #'   tmp <- withr::local_tempdir()
+  #'   withr::local_dir(tmp)    
+  #'   
+  #'   enable_local_models()
+  #'   enable_local_models(sub_folder=FALSE)
+  #' }
+  #' 
+  
   local_dir <- getwd()
   if (sub_folder) {
     local_dir <- file.path(local_dir, "pid_pos_models")
   }
-
+  
   if (!dir.exists(local_dir)) {
     if (!dir.create(local_dir, recursive = TRUE)) {
       stop("Could not create local model folder: ", local_dir)
     }
   }
-
+  
   set_model_folder(local_dir)
   invisible(local_dir)
 }
@@ -43,7 +49,16 @@ enable_package_models <- function() {
   #' @export
   #' @examples
   #' enable_package_models()
-  set_model_folder(user_data_dir("pid.pos"))
+  
+  cache_dir <- tools::R_user_dir("pid.pos", which = "cache")
+  
+  if (!dir.exists(cache_dir)) {
+    if (!dir.create(cache_dir, recursive = TRUE)) {
+      stop("Could not create local model folder: ", cache_dir)
+    }
+  }
+  set_model_folder(cache_dir)
+  
 }
 
 
@@ -62,17 +77,17 @@ set_udpipe_version <- function(version = c("2.5", "2.4", "2.3")) {
   if (!exists("pid.pos_env", envir = pkg_env)) {
     stop("pid.pos_env does not exist. Initialize it first.")
   }
-
+  
   version <- match.arg(version)
-
+  
   if (!version %in% names(pid.pos_env$allowed_repos)) {
     validation_error("No repository defined for version")
   }
-
+  
   # repo <- pid.pos_env$allowed_repos[[version]]
   # if (is.null(repo)) stop("No repository defined for version ", version)
   pid.pos_env$udpipe_version <- pid.pos_env$allowed_repos[[version]]
-
+  
   invisible(pid.pos_env$udpipe_version)
 }
 
@@ -81,10 +96,7 @@ summarize_repeated_sentences <- function(frm, ...) {
   #' @importFrom utils head read.csv write.csv
   first <- head(frm, 1)
   first$Repeats <- nrow(frm)
-  first$`Affected Columns` <- paste(
-    glue("`{unique(frm$Column)}`"),
-    collapse = ", "
-  )
+  first$`Affected Columns` <- paste(glue("`{unique(frm$Column)}`"), collapse = ", ")
   first
 }
 
