### 2026-02-20 def7dbd
Prevent unfound global note.
