+++ R/tag_data_frame.R
@@ -79,8 +79,8 @@ tag_data_frame <- function(frm,
     ungroup() |>
     mutate(ID = glue("Col:{Column} Row:{Row}")) |>
     dplyr::select(
-      .data$Document, .data$ID, .data$Repeats,
-      .data$`Affected Columns`, .data$PK
+      all_of(c("Document", "ID", "Repeats",
+      "Affected Columns", "PK"))
     )
 
   tag_frm <- tag_documents(
