+++ R/zzz.R
@@ -17,7 +17,7 @@ pid.pos_env <- new.env()
   )
 
   pid.pos_env$udpipe_repo <- pid.pos_env$allowed_repos[["2.5"]]
-  
+
   enable_package_models()
 
   invisible()
