+++ tests/testthat/test-utilities.R
@@ -10,24 +10,23 @@ test_that("set_model_folder", {
 test_that("creates and returns subfolder by default", {
   withr::local_dir(withr::local_tempdir())
   folder <- enable_local_models()
-  
+
   # Check folder was created
   expect_true(dir.exists(file.path(getwd(), "pid_pos_models")))
-  
+
   # Check returned path
   expect_equal(folder, file.path(getwd(), "pid_pos_models"))
-  
 })
 
 test_that("uses current working directory if sub_folder = FALSE", {
   # Create a temporary working directory
   withr::local_dir(withr::local_tempdir())
-  
+
   folder <- enable_local_models(sub_folder = FALSE)
-  
+
   # Should now point to the temp working directory
   expect_equal(folder, getwd())
-  
+
   # No pid_pos_models folder should exist
   expect_false(dir.exists(file.path(getwd(), "pid_pos_models")))
 })
@@ -35,40 +34,37 @@ test_that("uses current working directory if sub_folder = FALSE", {
 test_that("uses current working directory if sub_folder = FALSE", {
   withr::local_dir(withr::local_tempdir())
   folder <- enable_local_models(sub_folder = FALSE)
-  
+
   # Should be current working directory
   expect_equal(folder, getwd())
-  
+
   # No pid_pos_models folder should be created
   expect_false(dir.exists(file.path(getwd(), "pid_pos_models")))
-  
 })
 
 test_that("existing folder does not error", {
   withr::local_dir(withr::local_tempdir())
-  
+
   path <- file.path(getwd(), "pid_pos_models")
   dir.create(path)
-  
+
   # Should not error if folder exists
   folder <- enable_local_models()
   expect_equal(folder, path)
-  
 })
 
 test_that("returns folder path invisibly", {
   withr::local_dir(withr::local_tempdir())
   folder <- enable_local_models()
-  
+
   # Check return value is correct
   expect_equal(folder, file.path(getwd(), "pid_pos_models"))
-  
 })
 
 test_that("integration with set_model_folder sets correct path", {
   withr::with_tempdir({
     folder <- enable_local_models()
-    
+
     # Assuming get_model_folder() returns the currently set folder
     if (exists("get_model_folder")) {
       expect_equal(get_model_folder(), folder)
@@ -82,20 +78,19 @@ test_that("integration with set_model_folder sets correct path", {
 test_that("recursive folder creation works for nested subfolders", {
   withr::local_dir(withr::local_tempdir())
   nested_path <- file.path(getwd(), "a/b/c")
-  
+
   # Modify enable_local_models temporarily to create nested folder
   temp_fun <- function(sub_folder = TRUE) {
     local_dir <- nested_path
-    if (!dir.exists(local_dir))
+    if (!dir.exists(local_dir)) {
       dir.create(local_dir, recursive = TRUE)
+    }
     local_dir
   }
-  
+
   folder <- temp_fun()
   expect_true(dir.exists(nested_path))
   expect_equal(folder, nested_path)
-  
-  
 })
 
 # # Create a mock pid.pos_env for testing
@@ -105,13 +100,17 @@ test_that("recursive folder creation works for nested subfolders", {
 
 test_that("valid versions update environment correctly", {
   v <- set_udpipe_version("2.5")
-  expect_equal(pid.pos_env$udpipe_version,
-               "jwijffels/udpipe.models.ud.2.5")
+  expect_equal(
+    pid.pos_env$udpipe_version,
+    "jwijffels/udpipe.models.ud.2.5"
+  )
[…]
