### 2026-02-20 46beb32
Avoid overwriting temp dir for udpipe model in vignette
