### 2026-02-24 Stat-Cook/pid.pos#6 @Stat-Cook
@agerada Yeap - there's even a few helper functions (that may not be fully documented) to open the particular folder for the user to deposit the models.  I'm leaning towards two steps to make this the most seamless and tailorable:

1. Letting the user choose if they want to cache models or not
2. Letting the user choose between project level caching or package level (handled by `tools::R_user_dir("pid.pos", which = "cache")`)

I'm going to test to see if a user can pass in a `udpipe` object as the model, in theory it should pass through to the udpipe call.

On a broader remit - just finishing up the tools for custom taggers (including access to python nltk) and a redesign of the file crawler to make redacting a collection of files more seamless.  Next on the list is better documentation of creating a user specified redacter in line with one of your earliest suggestions.
