### 2026-02-25 #9728 @Stat-Cook
@adithirgis  Almost there with the code redesign - reviewer feedback proving very helpful.
