### 2026-02-25 d129457
Refactor to `cached` redact.
