### 2026-02-25 c67b051
Refactor read and process supported files
