### 2026-02-25 11771b0
Refactor redaction tools
