### 2026-02-25 9aa28d7
Refactor for efficiency
