+++ R/frame_replacement.R
@@ -1,101 +1,126 @@
-then_factory <- function(From, To, If) {
-  functions <- purrr::map2(
-    # From, To, function(.x, .y) function(vec) str_replace(vec, .x, .y)
-    From, To, function(.x, .y) function(vec) gsub(.x, .y, vec, fixed = T)
+#' @keywords internal
+if_function_factory <- function(df) {
+  pattern <- df$If[1]
+  
+  if (!all(df$If == pattern)) {
+    stop(sprintf("Rule block contains multiple 'If' values: %s", df$If))
+  }
+  
+  structure(\(vec) grepl(pattern, vec, fixed = TRUE),
+            class = "if_function",
+            If = pattern)
+}
+print.if_function <- function(x, ...) {
+  sprintf("`if_function` for %s", attr(x, "If")) |>
+    print()
+}
+
+#' @keywords internal
+then_function_factory <- function(from, to) {
+  structure(
+    \(vec) gsub(from, to, vec, fixed = TRUE),
+    class = "then_function",
+    from = from,
+    to = to
   )
-
-  function(vec) {
-    sapply(
-      vec,
-      \(i) functions |>
-        append(list(i), after = 0) |>
-        reduce(~ .y(.x))
-    )
-  }
+}
+print.then_function <- function(x, ...) {
+  sprintf("`then_function` for %s -> %s", attr(x, "from"), attr(x, "to")) |>
+    print()
 }
 
-if_factory <- function(If) {
-  function(vec) {
-    # str_detect(vec, If)
-    grepl(If, vec, fixed = T)
-  }
+#' @keywords internal
+then.list.factory <- function(df) {
+  purrr::map2(df$From, df$To, then.function.factory)
 }
 
-
-then.function.list <- function(rules.frm) {
-  #' @importFrom dplyr group_by group_map
-  rules.frm |>
-    group_by(.data$If) |>
-    group_map(
-      ~ then_factory(.x$From, .x$To, .y$If)
-    )
+#' @keywords internal
+rule.logic <- function(df) {
+  cond_fun <- if.function.factory(df)
+  replace_funs <- then.list.factory(df)
+  
+  list(condition = cond_fun, replace = replace_funs)
 }
 
 
-if.function.list <- function(rules.frm) {
-  rules.frm |>
-    group_by(.data$If) |>
-    group_map(
-      ~ if_factory(.y$If)
-    )
-}
-
+#' Replacement rules to redaction function
+#'
+#' Convert a `data.frame` of redaction rules into a function that can be applied to a character vector.
+#'
+#' @param rules.frm A data.frame with columns `If`, `From` and `To`.
+#'
+#' @importFrom purrr reduce map2
+#'
+#' @examples
+#' data(the_one_in_massapequa)
+#' example.data <- head(the_one_in_massapequa)
+#'
+#' raw_rules <- pid_pos(example.data) |>
+#'   report_to_redaction_rules()
+#'
+#' redaction_rules <- auto_replace(raw_rules,
+#'   replacement.f = random_replacement.f())
+#'
+#' redaction_func <- redaction_function_factory(redaction_rules)
+#'
+#' redaction_func(example.data)
+#'
+#' @export
 redaction_function_factory <- function(rules.frm) {
-  #' Create a redaction function from a `data.frame` of replacement rules.
-  #'
-  #' @param rules.frm A data.frame with columns `If`, `From` and `To`.
-  #'
-  #' @importFrom purrr reduce map2
-  then.functions <- then.function.list(rules.frm)
-  if.functions <- if.function.list(rules.frm)
-
-  function(vec) {
-    conditions <- map(if.functions, ~ .x(vec))
-
-    redactions <- map2(
-      conditions,
-      then.functions,
[…]
