### 2026-02-25 21b8a75
Change reader default method
