### 2026-02-25 4e5442e
Implement tests for base_error
