### 2026-02-25 6537cf8
Drop deprecated functionality
