+++ man/register_reader.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/register_reader.R
+\name{register_reader}
+\alias{register_reader}
+\title{Add a reader function for a specific file extension.}
+\usage{
+register_reader(fun, ext, env = pid.pos_env)
+}
+\arguments{
+\item{fun}{The function to read files with the specified extension. 
+The function should take a file path as its first argument and return a data frame.}
+
+\item{ext}{The file extension (e.g., "csv", "txt").}
+}
+\description{
+Add a reader function for a specific file extension.
+}
+\examples{
+# example code
+
+}
