+++ man/prepare_redactions.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/make_replacements.R
+% Please edit documentation in R/prepare_replacements.R
 \name{prepare_redactions}
 \alias{prepare_redactions}
 \title{Prepare a function from redaction rules.}
