+++ man/cached_redact_factory.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/efficient_redact_factory.R
-\name{efficient_redact_factory}
-\alias{efficient_redact_factory}
+% Please edit documentation in R/cached_redact_factory.R
+\name{cached_redact_factory}
+\alias{cached_redact_factory}
 \title{Stateful recoding template function}
 \usage{
-efficient_redact_factory(redact.function)
+cached_redact_factory(redact.function)
 }
 \arguments{
 \item{redact.function}{A single input function to perform variable recoding.}
