+++ man/redact.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/redact.R
+\name{redact}
+\alias{redact}
+\title{Redact PID}
+\usage{
+redact(object, redacter, in_batches = T, ...)
+}
+\arguments{
+\item{object}{The object to be redacted - either a vector or data frame}
+
+\item{redacter}{A `data.frame` of redaction rules or a function created by `redaction_function_factory()`.}
+}
+\description{
+Redact PID
+}
