+++ man/get_distinct_redaction_rules.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_distinct_redaction_rules.R
+\name{get_distinct_redaction_rules}
+\alias{get_distinct_redaction_rules}
+\title{Title}
+\usage{
+get_distinct_redaction_rules(object, include_context = FALSE)
+}
+\arguments{
+\item{object}{The object to extract distinct redaction rules from. 
+Can be a path to a folder of `pid` reports, a list of `pid` reports, or a single data frame.}
+
+\item{include_context}{A boolean flag indicating whether to include context information in the output. Default is FALSE.}
+}
+\description{
+Title
+}
