+++ man/batched_redact.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/efficient_redact_factory.R
-\name{efficient_redaction}
-\alias{efficient_redaction}
+% Please edit documentation in R/batched_redact.R
+\name{batched_redact}
+\alias{batched_redact}
 \title{A wrapper for efficient redaction.}
 \usage{
-efficient_redaction(frm, redact, n = NULL, .progress = T)
+batched_redact(frm, redact, n = NULL, .progress = T)
 }
 \arguments{
 \item{frm}{The data frame to be redacted}
