### 2026-02-25 a5917c0
Style code (GHA)
