+++ R/cached_redact_factory.R
@@ -1,12 +1,10 @@
 CachedRedact <- R6::R6Class("CachedRedact", list(
   redacted = character(),
   redact_function = NA,
-  
   initialize = function(redact_function) {
     self$redact_function <- redact_function
   },
   recode = function(vec) {
-   
     keys <- unique(vec[!is.na(vec)])
     uncoded <- self$get_uncoded_keys(keys)
 
@@ -20,15 +18,14 @@ CachedRedact <- R6::R6Class("CachedRedact", list(
     out <- self$redacted[vec]
     out[is.na(vec)] <- NA_character_
     unname(out)
-    
   },
   get_uncoded_keys = function(vec) {
     setdiff(vec, names(self$redacted))
   }
 ))
 
-#' @exportS3Method 
-print.CachedRedact <- function(x, ...){
+#' @exportS3Method
+print.CachedRedact <- function(x, ...) {
   sprintf("CachedRedact Object [size=%s]", length(x$redacted)) |>
     print()
 }
@@ -58,11 +55,11 @@ cached_redact_factory <- function(redact.function) {
   )
 }
 
-#' @exportS3Method 
-print.cached_redact_function <- function(x, ...){
-
-  sprintf("`cached_redact_function` [size=%s]", 
-          length(attr(x, "cache")$redacted)) |>
+#' @exportS3Method
+print.cached_redact_function <- function(x, ...) {
+  sprintf(
+    "`cached_redact_function` [size=%s]",
+    length(attr(x, "cache")$redacted)
+  ) |>
     print()
 }
-
