+++ R/read_data.R
@@ -7,7 +7,7 @@ read_delim <- function(path, sep, ...) {
     stringsAsFactors = FALSE,
     ...
   )
-  
+
   names(df) <- make.names(names(df), unique = TRUE)
   df
 }
@@ -15,13 +15,12 @@ read_delim <- function(path, sep, ...) {
 
 read_data <- function(file_path, ...) {
   ext <- tolower(tools::file_ext(file_path))
-  
+
   reader <- pid.pos_env$file_readers[[ext]]
-  
+
   if (is.null(reader)) {
     stop("Unsupported file type: ", ext, call. = FALSE)
   }
-  
+
   reader(file_path, ...)
 }
-
