+++ R/redact_internal.R
@@ -4,7 +4,7 @@ redact_internal <- function(object, redaction_func) {
 }
 
 redact_internal.data.frame <- function(object, redaction_func) {
-  #' @exportS3Method 
+  #' @exportS3Method
   object |>
     dplyr::mutate(
       dplyr::across(where(is.character), \(i) redaction_func(i))
@@ -12,7 +12,6 @@ redact_internal.data.frame <- function(object, redaction_func) {
 }
 
 redact_internal.default <- function(object, redaction_func) {
-  #' @exportS3Method 
+  #' @exportS3Method
   redaction_func(object)
 }
-
