### 2026-02-25 a224293
Implement test suite
