### 2026-02-25 cb91722
Expand test suite
