### 2026-02-25 604149d
Add error on col check
