### 2026-02-25 23ebe2c
Move functions to independent file
