+++ man/register_reader.Rd
@@ -7,7 +7,7 @@
 register_reader(fun, ext, env = pid.pos_env)
 }
 \arguments{
-\item{fun}{The function to read files with the specified extension. 
+\item{fun}{The function to read files with the specified extension.
 The function should take a file path as its first argument and return a data frame.}
 
 \item{ext}{The file extension (e.g., "csv", "txt").}
