+++ man/get_distinct_redaction_rules.Rd
@@ -7,7 +7,7 @@
 get_distinct_redaction_rules(object, include_context = FALSE)
 }
 \arguments{
-\item{object}{The object to extract distinct redaction rules from. 
+\item{object}{The object to extract distinct redaction rules from.
 Can be a path to a folder of `pid` reports, a list of `pid` reports, or a single data frame.}
 
 \item{include_context}{A boolean flag indicating whether to include context information in the output. Default is FALSE.}
