### 2026-02-25 c00dba3
Fix stubs
