### 2026-02-25 6522f30
refactor to redact_function
