### 2026-02-25 f86f7f3
refactor to redact_function
