### 2026-02-25 05789b7
Implement test suite
