+++ NAMESPACE
@@ -6,13 +6,13 @@ S3method(get_distinct_redaction_rules,character)
 S3method(get_distinct_redaction_rules,data.frame)
 S3method(get_distinct_redaction_rules,list)
 S3method(parse_redacter,data.frame)
-S3method(parse_redacter,redaction_function)
+S3method(parse_redacter,redact_function)
 S3method(prepare_redactions,character)
 S3method(prepare_redactions,data.frame)
 S3method(print,CachedRedact)
 S3method(print,cached_redact_function)
 S3method(print,if_function)
-S3method(print,redaction_function)
+S3method(print,redact_function)
 S3method(print,then_function)
 S3method(redact_internal,data.frame)
 S3method(redact_internal,default)
