### 2026-02-25 26b78a1
implement test suite
