### 2026-03-04 95fb8b6
Add inheritance
