+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method(batched_redact,default)
 S3method(get_distinct_redaction_rules,character)
 S3method(get_distinct_redaction_rules,data.frame)
 S3method(get_distinct_redaction_rules,list)
+S3method(parse_redacter,cached_redact_function)
 S3method(parse_redacter,data.frame)
 S3method(parse_redacter,redact_function)
 S3method(prepare_redactions,character)
