### 2026-03-11 8e98a77
Move functions to internal keyword
