+++ R/find_supported_files.R
@@ -4,10 +4,16 @@ get_implemented_extensions <- function() {
 }
 
 
+#' Read all supported files in a folder into a named list of data.frames
+#' 
+#' @param data_path  The file path at which data is stored
+#' @param extensions Optional.  The set of file extensions to scanned for.   
+#' @param verbose Boolean flag - if TRUE will...   
+#' 
+#' @keywords internal
 find_supported_files <- function(data_path,
                                  extensions = get_implemented_extensions(),
                                  verbose = FALSE) {
-  #' Read all supported files in a folder into a named list of data.frames
 
   if (!dir.exists(data_path)) {
     stop("data_path does not exist: ", data_path)
