+++ R/redact.R
@@ -3,6 +3,7 @@
 #' @param object The object to be redacted - either a vector or data frame
 #' @param redacter A `data.frame` of redaction rules or a function created by `redaction_function_factory()`.
 #'
+#' @export
 redact <- function(object, redacter, in_batches = T, ...) {
   redacter <- parse_redacter(redacter)
 
