### 2026-03-11 09395fc
init model folder in env
