### 2026-03-11 d3209f0
set model folder as tmp to avoid vestigial files
