+++ vignettes/AutoReplacement.Rmd
@@ -16,7 +16,17 @@ knitr::opts_chunk$set(
   rows.print = 6
 )
 library(pid.pos)
-pid.pos:::set_model_folder(tempdir())
+
+.tmp <- tempdir()
+pid.pos:::set_model_folder(.tmp)
+
+{
+  model <- "english-ewt-ud-2.5-191206.udpipe"
+  to <- file.path(.tmp, model)
+  url <- file.path("https://raw.githubusercontent.com/jwijffels/udpipe.models.ud.2.5/master/inst/udpipe-ud-2.5-191206/", model)
+  curl::curl_download(url, to, mode="wb")
+}
+
 ```
 
 This package is designed to support the identification and redaction of personally identifiable data (PID) within textual datasets.
@@ -42,7 +52,6 @@ library(pid.pos)
 
 report <- pid_pos(the_one_in_massapequa)
 replacement_rules <- report_to_redaction_rules(report)
-
 replacement_rules
 ```
 
@@ -79,20 +88,34 @@ updated_replacement_rules
 head(updated_replacement_rules, 5)
 ```
 
-The resulting `updated_replacement_rules` table can be converted into an executable redaction function using `prepare_redactions()`:
+The resulting `updated_replacement_rules`  can then be used alongside the original data in the `redact` function to adjust the data:
 
 ```{r}
-redact_function <- prepare_redactions(updated_replacement_rules)
+redact(
+  head(the_one_in_massapequa, 5),
+  updated_replacement_rules
+)
 ```
 
-This function can then be applied across character columns in the original dataset:
+If the quantity of text being redacted is large, and documents are regularly repeated, the user may wish 
+to parse the replacement rules into a caching redaction function:
 
-```{r}
-head(the_one_in_massapequa, 5) |>
-  dplyr::mutate(
-    dplyr::across(where(is.character), redact_function)
-  )
+``` {r}
+cached_redacter <- parse_redacter(updated_replacement_rules, with_cache = T)
+cached_redacter
 ```
+
+This new function has a `memoization` layer built in, so that if the same document is presented - replacements are called from memory.  This may speed up data processing if the same passage of text is presented multiple times, but comes at the cost of memory.  The `cached_redact_function` can be used in `redact` in the same way:
+
+``` {r}
+redacted_docs <-redact(
+  the_one_in_massapequa,
+  cached_redacter
+)
+cached_redacter
+```
+
+And its representation tracks the number of unique documents stored.
 
 ---
 
