### 2026-03-11 f6cb846
Redesign with 'friends' files
