+++ vignettes/FolderReport.Rmd
@@ -21,13 +21,22 @@ A common struggle in research is how data controllers can gain some level of con
 
 To help overcome these issues, as part of the `PID.POS` package we have implemented an API for the automated production  of proper noun reports on all files found within the same directory.  The intention is that should a collection of data sets be required for transfer, they can be placed in a single location, and the reports generated. 
 
-To demonstrate how this function works - we have supplied a collection of data sets featuring free text with the package.  The free text data draws on the `janeaustenr` package - constructing three csv files:
+To demonstrate how this function works - we have supplied a collection of data sets featuring free text with the package.  The free text data draws on the `friends` package - containing the scripts for three episodes of season 8:
 
-* `Emma.csv`
-* `PridePrejudice.csv`
-* `SenseSensability.csv`
+* `The One After 'I DO'.csv`
+* `The One Where Rachel Tells.csv`
+* `The One with the Red Sweater.csv`
 
-with each file consisting of 1000 rows and four columns (a primary key, a line of text from the book, a reference category, and the string-length of the text).  The first step in processing the data is to identify where the files are in your local folder structure:
+With each file consisting of 6 columns:
+
+With each file consisting of 6 columns:
+
+* `text` - The line of the script (dialogue or action)
+* `speaker` - The character uttering the line
+* `season` - Numeric ID of which season the episode was in
+* `episode` - Numeric ID of which episode the dialogue was in
+* `scene` 
+* `utterance`
 
 ```{r setup}
 library(pid.pos)
@@ -35,17 +44,18 @@ library(pid.pos)
 data_path <- system.file("vignette_data", package = "pid.pos")
 list.files(data_path)
 ```
+
 and we check the files are the intended data:
 
 ``` {r}
-emma.csv <- system.file("vignette_data", "Emma.csv", package = "pid.pos")
-kable(read.csv(emma.csv, nrows = 5))
+rachell_tells <- system.file("vignette_data", "The One Where Rachel Tells.csv", package = "pid.pos")
+rachell <- read.csv(rachell_tells, nrows = 5)
 ```
-To generate reports we call `report_on_folder` which takes three arguments:
+
+To generate reports we call `report_on_folder` which has two key arguments:
 
 * `data_path` - the path to the data directory
-* `report_dir` - a system path to where the proper noun reports should be saved
-* `to_remove` [optional] - a vector of columns to be ignored e.g. primary keys.
+* `report_dir` - [optional] a system path to where the proper noun reports should be saved
 
 ``` {r, eval=F}
 report_on_folder(data_path, report_dir = "Proper Noun Report")
@@ -57,14 +67,72 @@ Once evaluated the `report_dir` folder  gets populated by a set of csv files, on
 browseURL("Proper Noun Report")
 ```
 
-Each of these files consists of 5 columns:
+Each of these files consists of 6 columns:
 
-* `doc_id` - a reference of where the proper noun was detected
-* `token` - the proper noun detected
-* `sentence` - the full free text field
-* `Repeats` - how many times `sentence` appeared in the data set
-* `Affected Columns` - all the columns that `sentence` occured in.
+* `ID` - a reference of where the proper noun was detected
+* `Token` - the proper noun detected
+* `Sentence` - the sentence containing proper nouns
+* `Document` - the full document `Sentence` occurs in.
+* `Repeats` - how many times `Document` appeared in the data set
+* `Affected Columns` - all the columns that `Document` occurred in.
 
 ``` {r, eval=F}
-read.csv("Proper Noun Report/Emma.csv")
+read.csv("Proper Noun Report/The One Where Rachel Tells.csv")
+```
+
+The `report_on_folder` function is a high level API for the `pid_pos` package and 
[…]
