+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(get_distinct_redaction_rules,character)
 S3method(get_distinct_redaction_rules,data.frame)
 S3method(get_distinct_redaction_rules,list)
 S3method(parse_redacter,cached_redact_function)
+S3method(parse_redacter,character)
 S3method(parse_redacter,data.frame)
 S3method(parse_redacter,redact_function)
 S3method(prepare_redactions,character)
@@ -27,13 +28,14 @@ export(custom_tagger)
 export(enable_local_models)
 export(enable_package_models)
 export(filter_to_proper_nouns)
-export(frame_replacement)
 export(get_distinct_redaction_rules)
 export(hashing_replacement.f)
 export(merge_redactions)
+export(parse_redacter)
 export(pid_pos)
-export(prepare_redactions)
 export(random_replacement.f)
+export(redact)
+export(redact_at_folder)
 export(redaction_function_factory)
 export(register_reader)
 export(reinstate_default_reader)
@@ -41,7 +43,6 @@ export(report_on_folder)
 export(report_to_redaction_rules)
 export(set_udpipe_version)
 export(tag_data_frame)
-export(tag_documents)
 export(udpipe_factory)
 importFrom(R6,R6Class)
 importFrom(dplyr,across)
