+++ DESCRIPTION
@@ -21,7 +21,7 @@ Description: A package for the detection and contextualization of personally
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.3.1
+RoxygenNote: 7.3.3
 Imports:
     dplyr,
     udpipe,
@@ -42,7 +42,6 @@ Imports:
 Suggests: 
     knitr,
     rmarkdown,
-    openxlsx,
     testthat (>= 3.0.0),
     mockery,
     reticulate,
