+++ man/redact_supported_files.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/redact_supported_files.R
+\name{redact_supported_files}
+\alias{redact_supported_files}
+\title{Apply a redaction to a list of files.}
+\usage{
+redact_supported_files(
+  file_list,
+  output_path,
+  redacter,
+  export_function = NULL
+)
+}
+\arguments{
+\item{file_list}{File path to location of data sets}
+
+\item{output_path}{File path to write redacted data to}
+
+\item{redacter}{A redaction rules data frame or redaction function}
+
+\item{export_function}{A function to define export}
+}
+\description{
+Apply a redaction to a list of files.
+}
+\keyword{internal}
