+++ man/redact_at_folder.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/redact_at_folder.R
+\name{redact_at_folder}
+\alias{redact_at_folder}
+\title{API for redaction across a file structure}
+\usage{
+redact_at_folder(
+  data_path,
+  redacter,
+  output_path = "Redacted Data",
+  extensions = get_implemented_extensions(),
+  export_function = NULL,
+  verbose = FALSE
+)
+}
+\arguments{
+\item{data_path}{The file path at which data is stored}
+
+\item{redacter}{A redaction rules data frame or redaction function}
+
+\item{output_path}{File path to write redacted data to}
+
+\item{extensions}{Optional.  The set of file extensions to scanned for.}
+
+\item{export_function}{A function to define export}
+
+\item{verbose}{Boolean flag - if TRUE will...}
+}
+\description{
+API for redaction across a file structure
+}
