+++ R/merge_replacements.R
@@ -3,7 +3,7 @@ vector_merge_redactions <- function(vec,
                                     preprocess = utf8::utf8_encode) {
   #' @importFrom dplyr left_join join_by
   #'
-  #' @keywords internals 
+  #' @keywords internals
   # Bindings to suppress codetools warnings
   # about "no visible binding for global variable" `Old` and `If`
   Old <- If <- NA
@@ -24,7 +24,7 @@ merge_redactions <- function(frm,
   #' @param frm The data frame to be redacted
   #' @param cached_redactions A data frame with `If` and `Then` columns
   #' @param preprocess A function of preprocessing steps to be applied to the text columns.
-  #' 
+  #'
   #' @export
   frm |>
     mutate(dplyr::across(
