### 2026-03-11 3f60b96
Mock vignette reports for kniting
