### 2026-03-11 4dc080f
Update docs for internals with explicit pid.pos call
