+++ R/file_not_found_error.R
@@ -12,7 +12,7 @@
 #' @examples
 #' #' # Example of using type_error
 #' f <- function() {
-#'   file_not_found_error("This is a type error", call = rlang::caller_env())
+#'   pid.pos:::file_not_found_error("This is a type error", call = rlang::caller_env())
 #' }
 #'
 #' # Trigger the error safely without stopping R
