+++ man/prepare_redactions.Rd
@@ -30,7 +30,7 @@ redactions.raw <- report_to_redaction_rules(report)
 replace_by <- random_replacement.f()
 redactions <- auto_replace(redactions.raw, replacement.f = replace_by)
 
-f <- prepare_redactions(redactions)
+f <- pid.pos:::prepare_redactions(redactions)
 f(example.data$text)
 }
 
