+++ man/type_error.Rd
@@ -22,7 +22,7 @@ A custom abort function
 \examples{
 #' # Example of using type_error
 f <- function() {
-  type_error("This is a type error")
+  pid.pos:::type_error("This is a type error")
 }
 
 # Trigger the error safely without stopping R
