+++ DESCRIPTION
@@ -34,6 +34,7 @@ Imports:
     glue,
     progress,
     readxl,
+    writexl,
     stringr,
     utf8,
     tidyr,
