+++ DESCRIPTION
@@ -34,7 +34,6 @@ Imports:
     glue,
     progress,
     readxl,
-    writexl,
     stringr,
     utf8,
     tidyr,
@@ -47,6 +46,8 @@ Suggests:
     mockery,
     reticulate,
     covr,
+    curl,
+    writexl,
     withr
 Config/testthat/edition: 3
 VignetteBuilder: knitr
