### 2026-03-11 9d11bfa
update docstrings
