+++ man/get_context.Rd
@@ -11,6 +11,8 @@ get_context(
 )
 }
 \arguments{
+\item{context_window}{The width of window around the token to be taken.}
+
 \item{sentences}{A character vector of sentences.}
 
 \item{tokens}{A character vector of tokens.}
