+++ R/get_context.R
@@ -2,8 +2,8 @@
 #'
 #' NB: to set the context window size, use `set_context_window()`.
 #'
-#' @param sentences A character vector of sentences.
-#' @param tokens A character vector of tokens.
+#' @param sentence A character vector of sentences.
+#' @param token A character vector of tokens.
 #' @param context_window The width of window around the token to be taken.
 #'
 #' @importFrom dplyr mutate
