+++ R/report_on_folder.R
@@ -15,6 +15,7 @@ empty.report <- data.frame(
 #' @inheritParams pid_pos
 #' @param export_function A function to control exporting the reports to disk.  Current 
 #'  options  are `export_as_tree` and `export_flat`
+#' @param report_dir The location to write PID reports to
 #'
 #' @return NULL
 #'
