### 2026-03-11 7141574
Remove environment arg
