+++ R/register_reader.R
@@ -1,4 +1,4 @@
-register_reader <- function(fun, ext, env = pid.pos_env) {
+register_reader <- function(fun, ext) {
   #' Add a reader function for a specific file extension.
   #'
   #' @param fun The function to read files with the specified extension.
@@ -17,12 +17,12 @@ register_reader <- function(fun, ext, env = pid.pos_env) {
     stop("The reader must be a function.", call. = FALSE)
   }
 
-  env$file_readers[[tolower(ext)]] <- fun
+  pid.pos_env$file_readers[[tolower(ext)]] <- fun
   invisible(TRUE)
 }
 
 
-reinstate_default_reader <- function(env = pid.pos_env) {
+reinstate_default_reader <- function() {
   #' Reinstate the default read functionality for csv, tsv, xls, and xlsx files.
   #'
   #' @export
@@ -37,5 +37,5 @@ reinstate_default_reader <- function(env = pid.pos_env) {
     xls = readxl::read_excel,
     xlsx = readxl::read_excel
   )
-  invisible(purrr::imap(default_readers, \(.x, .y) register_reader(.x, .y, env = env)))
+  invisible(purrr::imap(default_readers, \(.x, .y) register_reader(.x, .y)))
 }
