### 2026-03-11 cfa1eaf
Implement redact test suite
