+++ tests/testthat/test-redact.R
@@ -1,39 +1,38 @@
 # # Example input
 # vec <- c("secret info", "public info")
 # df <- data.frame(a = vec, b = vec)
-# 
+#
 # parsed_reacter_mock <- function(redacter) "parsed_redacter"
 # redact_internal_mock <- function(...) "internal_result"
 # batched_redact_mock <- function(object, redacter, ...) "batched_result"
-# 
+#
 # test_that("redact calls redact_internal when in_batches = FALSE", {
 #   mockery::stub(redact, "parse_redacter", parsed_reacter_mock)
 #   mockery::stub(redact, "redact_internal", redact_internal_mock)
-#   
+#
 #   result <- redact(vec, redacter = "dummy", in_batches = FALSE)
-#   
+#
 #   expect_equal(result, "internal_result")
 # })
-# 
+#
 # test_that("redact calls batched_redact when in_batches = TRUE", {
 #   mockery::stub(redact, "parse_redacter", parsed_reacter_mock)
 #   mockery::stub(redact, "batched_redact", batched_redact_mock)
-#   
+#
 #   result <- redact(vec, redacter = "dummy", in_batches = TRUE)
-#   
+#
 #   expect_equal(result, "batched_result")
 # })
-# 
+#
 # test_that("parse_redacter is called with the provided redacter", {
 #   called_with <- NULL
 #   mockery::stub(redact, "parse_redacter", function(redacter) { called_with <<- redacter; parsed_reacter_mock(redacter)})
 #   mockery::stub(redact, "redact_internal", redact_internal_mock)
-#   
+#
 #   redact(vec, redacter = "my_rules", in_batches = FALSE)
-#   
+#
 #   expect_equal(called_with, "my_rules")
 # })
-
 
 
 vec <- c("secret info", "public info")
@@ -47,7 +46,7 @@ batched_redact_mock <- function(object, redacter, ...) "batched_result"
 test_that("redact calls redact_internal when in_batches = FALSE", {
   mockery::stub(redact, "parse_redacter", parsed_reacter_mock)
   mockery::stub(redact, "redact_internal", redact_internal_mock)
-  
+
   result <- redact(vec, redacter = "dummy", in_batches = FALSE)
   expect_equal(result, "internal_result")
 })
@@ -55,7 +54,7 @@ test_that("redact calls redact_internal when in_batches = FALSE", {
 test_that("redact calls batched_redact when in_batches = TRUE", {
   mockery::stub(redact, "parse_redacter", parsed_reacter_mock)
   mockery::stub(redact, "batched_redact", batched_redact_mock)
-  
+
   result <- redact(vec, redacter = "dummy", in_batches = TRUE)
   expect_equal(result, "batched_result")
 })
@@ -67,7 +66,7 @@ test_that("parse_redacter is called with the provided redacter", {
     parsed_reacter_mock(redacter)
   })
   mockery::stub(redact, "redact_internal", redact_internal_mock)
-  
+
   redact(vec, redacter = "my_rules", in_batches = FALSE)
   expect_equal(called_with, "my_rules")
 })
@@ -102,7 +101,7 @@ redact_internal_real <- function(object, redacter, ...) {
 test_that("redact works on character vector realistically", {
   mockery::stub(redact, "parse_redacter", function(redacter) simple_redacter)
   mockery::stub(redact, "redact_internal", redact_internal_real)
-  
+
   result <- redact(vec, redacter = simple_redacter, in_batches = FALSE)
   expect_equal(result, c("*** info", "public info"))
 })
@@ -110,7 +109,7 @@ test_that("redact works on character vector realistically", {
 test_that("redact works on data frame realistically", {
   mockery::stub(redact, "parse_redacter", function(redacter) simple_redacter)
[…]
