### 2026-03-11 25e98bd
Extend class types
