### 2026-03-11 ac048ce
Add defensive layers
