### 2026-03-11 674612d
Refactor our `prepare_redactions`
