### 2026-03-11 8138ae0
Refactor deprecated functions.
