+++ R/find_supported_files.R
@@ -17,7 +17,7 @@ find_supported_files <- function(data_path,
   if (!is.character(data_path) | length(data_path) != 1) {
     stop("`data_path` should be a single string")
   }
-  
+
   if (!dir.exists(data_path)) {
     stop("data_path does not exist: ", data_path)
   }
