### 2026-03-11 fd2774b
Deprecate `read_delim`
