+++ R/read_data.R
@@ -1,17 +1,3 @@
-read_delim <- function(path, sep, ...) {
-  df <- utils::read.table(
-    path,
-    sep = sep,
-    header = TRUE,
-    check.names = FALSE,
-    stringsAsFactors = FALSE,
-    ...
-  )
-
-  names(df) <- make.names(names(df), unique = TRUE)
-  df
-}
-
 
 read_data <- function(file_path, ...) {
   ext <- tolower(tools::file_ext(file_path))
