### 2026-03-11 1a8eebd
inherit error into custom error
