### 2026-03-11 19ae2a2
Test error behaviours
