### 2026-03-11 b40b9b7
Add zzz.R to the ignore list in codecov.yml
