### 2026-03-11 7b88e7f
Add workflow_dispatch trigger to test coverage
