+++ .github/workflows/test-coverage.yaml
@@ -1,6 +1,7 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
+  workflow_dispatch:
   push:
     branches: [main, master, redesign, JOSSReviewRecommendations]
     paths: ["**.R", "DESCRIPTION", "NAMESPACE"]
