### 2026-03-11 96c0c26
Add workflow_dispatch trigger to test coverage workflow
