### 2026-03-11 d220e85
Update ignore pattern for zzz.R in codecov.yml
