+++ codecov.yml
@@ -19,4 +19,4 @@ coverage:
 ignore:
   - "**/browse.R"
   - "browse.R"
-  - "zzz.R"
+  - "**/zzz.R"
