+++ codecov.yml
@@ -20,3 +20,4 @@ ignore:
   - "**/browse.R"
   - "browse.R"
   - "**/zzz.R"
+  - "zzz.R"
