### 2026-03-13 3032438
Add default method for parse_redacter
