+++ NAMESPACE
@@ -8,6 +8,7 @@ S3method(get_distinct_redaction_rules,list)
 S3method(parse_redacter,cached_redact_function)
 S3method(parse_redacter,character)
 S3method(parse_redacter,data.frame)
+S3method(parse_redacter,default)
 S3method(parse_redacter,redact_function)
 S3method(prepare_redactions,character)
 S3method(prepare_redactions,data.frame)
