+++ R/udpipe_factory.R
@@ -4,10 +4,13 @@
 #' The returned function accepts a character vector of documents and returns
 #' a tibble with tokens, sentences, and token metadata.
 #'
-#' @param model Character. The name of the UDPipe model to use. Defaults to `"english-ewt"`.
-#' @param model_dir Character. Directory where UDPipe models are stored. Defaults to `pid.pos_env$model_folder`.
-#' @param udpipe_repo Character. URL or path of the UDPipe model repository. Defaults to `pid.pos_env$udpipe_version`.
-#'
+#' @param model Character. The name of the UDPipe model to use. Defaults to `english-ewt`.
+#' @param model_dir Character. Directory where UDPipe models are stored. 
+#' @param udpipe_repo Character. URL or path of the UDPipe model repository. 
+#' 
+#' @seealso [enable_local_models()], [enable_package_models()] and  
+#'   [set_udpipe_version()] for control of the configuration environment.
+#' 
 #' @return A function that takes a character vector of documents and returns a `tibble`
 #' with columns:
 #' \describe{
@@ -16,7 +19,8 @@
 #'   \item{Sentence}{Sentence containing the token}
 #'   \item{upos}{The universal parts of speech tag of the token. See https://universaldependencies.org/format.html}
 #' }
-#' and all columns returned by the `udpipe::udpipe()` function for each token.
+#' and all columns returned by the <[udpipe()`][udpipe::udpipe]> 
+#'   function for each token.
 #'
 #' @export
 #'
