### 2026-03-13 03e5440
Reformat doc strings
