+++ .github/workflows/pkgdown.yaml
@@ -1,8 +1,9 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/v2/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
+  workflow_dispatch:
   push:
-    branches: [main, master, redesign]
+    branches: [main, master, redesign, JOSSReviewRecommendations]
     paths: ["**.R", "README.md", "**.Rmd", "**.Rd"]
   pull_request:
   release:
