+++ .github/workflows/pkgdown.yaml
@@ -8,7 +8,6 @@ on:
   pull_request:
   release:
     types: [published]
-  workflow_dispatch:
 
 name: pkgdown.yaml
 
