+++ reference/enable_package_models.md
@@ -1,0 +1,21 @@
+# Set the model folder to the package data folder.
+
+Intended if you want to share udpipe models between different R
+projects.
+
+## Usage
+
+``` r
+enable_package_models()
+```
+
+## Value
+
+The path to the model folder.
+
+## Examples
+
+``` r
+enable_package_models()
+#> [1] "/home/runner/.cache/R/pid.pos"
+```
