+++ reference/set_udpipe_version.md
@@ -1,0 +1,26 @@
+# Set the udpipe model repository version.
+
+Set the udpipe model repository version.
+
+## Usage
+
+``` r
+set_udpipe_version(version = c("2.5", "2.4", "2.3"))
+```
+
+## Arguments
+
+- version:
+
+  Character. The udpipe model version to use. One of "2.5", "2.4", or
+  "2.3".
+
+## Value
+
+Character. The udpipe model repository.
+
+## Examples
+
+``` r
+set_udpipe_version("2.4")
+```
