+++ articles/folder-report.md
@@ -1,0 +1,958 @@
+# Reporting on every file in a folder
+
+## The `PID.POS` report
+
+A common struggle in research is how data controllers can gain some
+level of confidence that large data sets don’t contain personally
+identifiable data. In some cases, this job only requires a brief
+inspection and columns that often contain PID such as name, or ID are
+obvious. However, data sets can contain broad free text fields, fields
+that are only needed in a small number of cases, or may have been
+shifted - placing PID in harder to detect locations. If the data sets
+consist of ~10,000 or more observations manual inspection of rare PID
+has only a limited chance of finding problems, let alone the resource
+cost required to do any pass of the data.
+
+To help overcome these issues, as part of the `PID.POS` package we have
+implemented an API for the automated production of proper noun reports
+on all files found within the same directory. The intention is that
+should a collection of data sets be required for transfer, they can be
+placed in a single location, and the reports generated.
+
+To demonstrate how this function works - we have supplied a collection
+of data sets featuring free text with the package. The free text data
+draws on the `friends` package - containing the scripts for three
+episodes of season 8:
+
+- `The One After 'I DO'.csv`
+- `The One Where Rachel Tells.csv`
+- `The One with the Red Sweater.csv`
+
+With each file consisting of 6 columns:
+
+With each file consisting of 6 columns:
+
+- `text` - The line of the script (dialogue or action)
+- `speaker` - The character uttering the line
+- `season` - Numeric ID of which season the episode was in
+- `episode` - Numeric ID of which episode the dialogue was in
+- `scene`
+- `utterance`
+
+``` r
+library(pid.pos)
+
+data_path <- system.file("vignette_data", package = "pid.pos")
+list.files(data_path)
+#> [1] "The_One_After_I_Do.csv"           "The_One_Where_Rachel_Tells.csv"  
+#> [3] "The_One_with_the_Red_Sweater.csv"
+```
+
+and we check the files are the intended data:
+
+``` r
+rachell_tells <- system.file("vignette_data", "The_One_Where_Rachel_Tells.csv", package = "pid.pos")
+rachell <- read.csv(rachell_tells, nrows = 5)
+```
+
+To generate reports we call `report_on_folder` which has two key
+arguments:
+
+- `data_path` - the path to the data directory
+- `report_dir` - \[optional\] a system path to where the proper noun
+  reports should be saved
+
+``` r
+report_on_folder(data_path, report_dir = "Proper Noun Report")
+```
+
+Once evaluated the `report_dir` folder gets populated by a set of csv
+files, one per data set found at `data_path`:
+
+``` r
+browseURL("Proper Noun Report")
+```
+
+Each of these files consists of 6 columns:
+
+- `ID` - a reference of where the proper noun was detected
+- `Token` - the proper noun detected
+- `Sentence` - the sentence containing proper nouns
+- `Document` - the full document `Sentence` occurs in.
+- `Repeats` - how many times `Document` appeared in the data set
+- `Affected Columns` - all the columns that `Document` occurred in.
+
+``` r
+read.csv("Proper Noun Report/The_One_Where_Rachel_Tells.csv")
+```
+
+    #>                      ID     Token
+    #> 1        Col:text Row:1    Monica
+    #> 2        Col:text Row:1  Chandler
+    #> 3        Col:text Row:1    Monica
+    #> 4        Col:text Row:1  Chandler
+    #> 5        Col:text Row:1    Monica
+    #> 6        Col:text Row:2      Babe
[…]
