+++ R/pid_pos.R
@@ -5,18 +5,18 @@
 #' proper noun, and how often the `Document` occurs.
 #'
 #' @param frm A data frame to check for proper nouns
-#' @param tagger [optional] Either a string naming a UDPipe model (see  \link[udpipe]{udpipe_download_model} 
-#'   for the list of models) or a custom tagging function (see \code{vignette("custom-functions")}
+#' @param tagger   Either a string naming a UDPipe model (see  [udpipe::udpipe_download_model] 
+#'   for the list of models) or a custom tagging function (see [Custom Functions](../articles/custom-functions.html)
 #'   for details of what is required).
-#' @param filter_func [optional] A function to filter the tagged instances. See
-#'   the 'Custom Filtering Functions' section of \code{vignette("custom-functions")}
+#' @param filter_func   A function to filter the tagged instances. See
+#'   the 'Custom Filtering Functions' section of [Custom Functions](../articles/custom-functions.html
 #'   for more details.   
-#' @param chunk_size [optional] The number of sentences to tag at a time.  The optimal value
+#' @param chunk_size   The number of sentences to tag at a time.  The optimal value
 #'   has yet to be determined.
-#' @param to_ignore [optional] A vector of column names to be ignored by the algorithm.
+#' @param to_ignore   A vector of column names to be ignored by the algorithm.
 #'   Intended to be used for variables that are giving strong false positives, such as
 #'   IDs or ICD-10 codes.
-#' @param warn_if_missing [optional] Raise a warning if the `to_ignore` columns are
+#' @param warn_if_missing   Raise a warning if the `to_ignore` columns are
 #'   not in the data frame.
 #'
 #' @return A `pid_report` (inheriting from tibble) containing:
