+++ R/report_to_redaction_rules.R
@@ -5,8 +5,8 @@
 #' redactions.
 #'
 #' @param report The `data.frame` generated by `pid_pos`
-#' @param path [optional] File path for write.
-#' @param include_context [optional] Binary-flag. If True, adds a `Context` column giving a subset
+#' @param path  File path for write.
+#' @param include_context  Binary-flag. If True, adds a `Context` column giving a subset
 #' of the sentence being redacted around the token.
 #'
 #' @return A tibble with columns:
