+++ man/base_error.Rd
@@ -11,7 +11,7 @@ base_error(subclass, message, ..., call)
 
 \item{message}{The error message to display}
 
-\item{...}{Additional arguments to pass to `abort()`}
+\item{...}{Additional arguments to pass to \code{abort()}}
 
 \item{call}{The call environment to use for the error (defaults to the caller's environment)}
 }
