+++ reference/report_to_redaction_rules.md
@@ -1,8 +1,7 @@
 # Initialize redaction rules
 
-Convert the output of \`pid_pos\` to a csv file for editing. It is
-intended the user changes the \`To\` column to define desired
-redactions.
+Convert the output of `pid_pos` to a csv file for editing. It is
+intended the user changes the `To` column to define desired redactions.
 
 ## Usage
 
@@ -14,27 +13,27 @@ report_to_redaction_rules(report, path = NULL, include_context = FALSE)
 
 - report:
 
-  The \`data.frame\` generated by \`pid_pos\`
+  The `data.frame` generated by `pid_pos`
 
 - path:
 
-  \[optional\] File path for write.
+  File path for write.
 
 - include_context:
 
-  \[optional\] Binary-flag. If True, adds a \`Context\` column giving a
-  subset of the sentence being redacted around the token.
+  Binary-flag. If True, adds a `Context` column giving a subset of the
+  sentence being redacted around the token.
 
 ## Value
 
 A tibble with columns:
 
-- \`If\`: The condition (sentence) under which the replacement is to be
+- `If`: The condition (sentence) under which the replacement is to be
   made.
 
-- \`From\`: The original token.
+- `From`: The original token.
 
-- \`To\`: The replacement token.
+- `To`: The replacement token.
 
 ## Examples
 
