### 2026-04-16 Stat-Cook/pid.pos#6 @Stat-Cook
@agerada We've now made significant additions to the package and reworked the paper to better reflect all the functionality included.  For the purposes of controlling downloads - on first use the user should now be asked for explicit permission to fetch models in line with CRAN guidelines with the option in `pidpos_setup()` to force environemnt only models (with an example).

There is also a vignette dedicated to guiding users in extending the package with their own functionality (`custom-functions`).  Any suggestions/ advice/ comments most welcome.
