### 2026-04-29 Stat-Cook/pid.pos#6 @agerada
Hi @Stat-Cook.

I have now had some time to review the updated package. The internal functions have become a bit more complicated, for example the package uses custom errors and warnings; feels a bit overkill for a package of this size, but this is more of a design choice preference.

The user-facing functions otherwise seem similar. I note the functionality to use local models. I haven't tested this myself just yet, and can't see a unit test for it.

My main comment at this point is regarding the package's main use of detecting PID. As discussed above, the package uses the UDPipe model to find proper nouns. Unfortunately I do not think this suffices for reasonable detection of PID. For example, when using an example file from `NLM Scrubber` (another software that removes PID), I get the following results (using the file `dummy_scrubber_test_1.txt` from https://lhncbc.nlm.nih.gov/scrubber/download.html):

```
# input text:
 1 Date of Consult:
 2 April 16, 2019
 3 Reason for Consult:
 4 Sleep apnea
 5 History of Present Illness:
 6 John Smith is a 65 year old man who presented to his physician's office with his wife, Jane Smith. Jane wanted John to come in for a…
 7 Family History:
 8 John's mother, Jill, is a 91 year old woman with a history of sleep apnea. John's father, Jack, died at age 84, who died in a car ac…
 9 Consult:
10 See full report in HR19-0000000A
11 Plan:
12 Admit patient to Sleep Study Protocol #00-AA-0000. John is scheduled for a polysomnography for April 29, 2019 in Tampa, Florida.
13 Please send full report to:
14 Dr. Robert Jones
15 123 Main St., Suite #333
16 Tampa, Florida 33602

# after redaction using pidpos:
 1 Date of Consult:
 2 FDLVFWIUVL 16, 2019
 3 TBLHAWFEUK for XFUSGOOTGV:
 4 Sleep apnea
 5 History of SPJOMTLMOO GJMQCEJUHF:
 6 MNGDQZWFTZ Smith is a 65 year old man who presented to his physician's office with his wife, CGJBNLVRUB Smith. CGJBNLVRUB wanted MNG…
 7 Family History:
 8 MNGDQZWFTZ's mother, Jill, is a 91 year old woman with a history of sleep apnea. MNGDQZWFTZ's father, SRWMLYPZMV, died at age 84, wh…
 9 Consult:
10 See full report in HR19-0000000A
11 Plan:
12 Admit patient to ZWHWXDHUNX LWYRHQXDTL DCZIOHMVZP #00-AA-0000. MNGDQZWFTZ is scheduled for a polysomnography for FDLVFWIUVL 29, 2019…
13 Please send full report to:
14 WTZSGGIWMN GRFMRWABUW QIRAEZLBMV
15 123 Main QJKULRKSTS, QWRXFEDOJP #333
16 YQOVUTITWF, XVPYJTCXKG 33602
```

As you can see, `pidpos` misses names (Smith, Jill). It is also not designed to remove dates or ID numbers. Text is generally still regarded as identifable while these are present.

Furthermore, if the text is first converted to lowercase, then the misses are amplified:

```
 1 date of consult:
 2 april 16, 2019
 3 reason for consult:
 4 sleep apnea
 5 history of present illness:
 6 PJSYDJSPFC smith is a 65 year old man who presented to his physician's office with his wife, jane smith. jane wanted PJSYDJSPFC to c…
 7 family history:
 8 john's mother, jill, is a 91 year old woman with a history of sleep apnea. john's father, CXGYAAIOPQ, died at age 84, who died in a …
 9 consult:
10 see full report in hr19-0000000a
11 plan:
12 admit patient to sleep study protocol #00-aa-0000. john is scheduled for a polysomnography for april 29, 2019 in CXVUGIWNHY, MYONUDG…
13 please send full report to:
14 dr. robert jones
15 123 main st., suite #333
16 tampa, MYONUDGUDV 33602
```

I appreciate that it could be argued that these are limitations of the model being used, but the claim is that `pidpos` detects PID, then it has to be scrutinised as such. The software does provide an API that allows for custom tagging functions; realistically, by that point most users would switch to a software that also detects dates, lowercase proper nouns, etc.

In summary, my main concern at present is that the software remains predominantly a wrapper around `UDPipe`. I commented on this in my initial review, but on testing further with testcases (such as the above), the limitations are more evident. Please let me know if your experience has been different (when testing on other text).

For the software to be useful for the detection of PID, it needs a more robust model (or approach) that captures text such as lowercase proper nouns, ID numbers, dates, and more. Ideally it should be validated with testing text (either publicly available, or, given the difficulties of obtaining text with PID, at least synthetically-generated).
