### 2026-04-30 Stat-Cook/pid.pos#6 @Stat-Cook
Hi @agerada  - very insighful thank you.  The custom errors were our solution to the extendability issues actually so that tryCatch blocks can dispatch to different paths depending on the error logic.  Possibly overboard!

 The lower case issue here is interesting -  we've decided to explore how easily we can include a BERT model to improve performance but keep the current replacement and reporting infrastructure.   Once we have a set of models we'll add a vignette to evidence how they perform in each scenario and hopefully that'll give users a solid choice.
