### 2026-04-30 Stat-Cook/pid.pos#6 @agerada
Hi @Stat-Cook. Deidentification of free text and unstructured data is a difficult problem, and any advance will be welcome. In the JOSS paper, the `pii` package is referenced. Interestingly, this package is actually for the detection of identifying data in _structured_ data, so it would be useful alongside `pidpos`.

My hunch is that the use of parts of speech alone won't be sufficient (as seen in the example above), so I agree with looking at other models such as BERT. Almost certainly multiple concurrent approaches will be needed. You may already be aware of these, but the following may be useful for validation:

https://www.physionet.org/content/deid/1.1/

and https://www.physionet.org/content/deidentifiedmedicaltext/1.0/

Unfortunately, the medical text is credentialed access (not open access), but would be helpful to understand performance. I believe this software was used to deidentify the MIMIC-IV dataset.
