### 2026-04-30 Stat-Cook/pid.pos#6 @Stat-Cook
@agerada  Really helpful thanks - the competing priority is also the usability with minimal blockage in the NHS.  There are tools like [NLM scrubber](https://lhncbc.nlm.nih.gov/scrubber/index.html) but it means a java runtime and hence multiple IT redflags.  The problem persists for many of the easily available tools (python is also pretty restricted in our experience) which is why udpipe (with distinct limitations) has been the central point - it's an interesting balancing act of whats useful in the most open context with whats useful in a very closed context!
